Test pathway association with AES-PCA
Source: R/aesPC_wrapper.R

Given a supervised OmicsPath object (one of OmicsSurv, OmicsReg, or OmicsCateg), extract the first \(k\) adaptive, elastic-net, sparse principal components (PCs) from each expressed pathway in the -Omics assay design matrix, test their association with the response matrix, and return a data frame of the adjusted \(p\)-values for each pathway.

AESPCA_pVals(object, numPCs = 1, min.features = 3, numReps = 1000,
  parallel = FALSE, numCores = NULL, adjustpValues = TRUE,
  adjustment = c("Bonferroni", "Holm", "Hochberg", "SidakSS", "SidakSD", "BH",
  "BY", "ABH", "TSBH"), ...)

# S4 method for OmicsPathway
AESPCA_pVals(object, numPCs = 1, min.features = 3,
  numReps = 1000, parallel = FALSE, numCores = NULL,
  adjustpValues = TRUE, adjustment = c("Bonferroni", "Holm", "Hochberg",
  "SidakSS", "SidakSD", "BH", "BY", "ABH", "TSBH"), ...)

Arguments

object
An object of class OmicsPathway with a response matrix or vector.

numPCs
The number of PCs to extract from each pathway. Defaults to 1.

min.features
What is the smallest number of genes allowed in each pathway? This argument must be kept constant across all calls to this function which use the same pathway list. Defaults to 3.

numReps
The number of permutations to take of the data to calculate a \(p\)-value for each pathway. Defaults to 1000.

parallel
Should the computation be completed in parallel? Defaults to FALSE.

numCores
If parallel = TRUE, how many cores should be used for computation? Defaults to NULL.

adjustpValues
Should you adjust the \(p\)-values for multiple comparisons? Defaults to TRUE.

adjustment
Character vector of procedures. The returned data frame will be sorted in ascending order by the first procedure in this vector, with ties broken by the unadjusted \(p\)-value. If only one procedure is selected, then it is necessarily the first procedure. See the documentation for the adjustRaw_pVals function for the adjustment procedure definitions and citations.

...
Dots for additional internal arguments.

Value

A data frame with columns:

pathways : The names of the pathways in the Omics* object (given in object@pathwaySet$pathways.)

setsize : The number of genes in each of the original pathways (given in the object@pathwaySet$setsize object).

terms : The pathway description, as given in the object@pathwaySet$TERMS object.

rawp : The unadjusted \(p\)-values of each pathway.

... : Additional columns as specified through the adjustment argument.

Some of the pathways in the supplied pathway set list will be removed, or "trimmed", during function execution. These trimmed pathways will have \(p\)-values given as NA. For an explanation of pathway trimming, see the documentation for the expressedOmes function.

The data frame will be sorted in ascending order by the method specified first in the adjustment argument. If adjustpValues = FALSE, then the data frame will be sorted by the raw \(p\)-values. If you have the suggested tidyverse package suite loaded, then this data frame will print as a tibble. Otherwise, it will print as a data frame.

Details

This is a wrapper function for the expressedOmes, extract_aesPCs, permTest_OmicsSurv, permTest_OmicsReg, and permTest_OmicsCateg functions.

See also

Examples

# NOT RUN {
  ###  Load the Example Data  ###
  data("colonSurv_df")
  data("colon_pathwaySet")

  ###  Create an OmicsSurv Object  ###
  colon_OmicsSurv <- create_OmicsSurv(assayData_df = colonSurv_df[, -(1:2)],
                                      pathwaySet_ls = colon_pathwaySet,
                                      eventTime_num = colonSurv_df$OS_time,
                                      eventObserved_lgl = as.logical(colonSurv_df$OS_event))

  ###  Calculate Pathway p-Values  ###
  colonSurv_pVals_df <- AESPCA_pVals(object = colon_OmicsSurv,
                                     numReps = 500,
                                     parallel = TRUE,
                                     numCores = 2,
                                     adjustpValues = TRUE,
                                     adjustment = c("Hoch", "SidakSD"))
# }