Check an Input Assay

Check the classes, dimensions, missingness, feature variance, feature type, and feature names of a data frame.

CheckAssay(df, removeNear0 = TRUE, epsilon = 10^-6)

Arguments

df
An assay data frame supplied to the CreateOmics function. The first column is assumed to be the sample IDs, and will be ignored. See CheckSampleIDs for checking sample IDs.

removeNear0
Should columns of df with variance near 0 be removed? Defaults to TRUE.

epsilon
Threshold to consider the variance of a column equal to 0. Defaults to 0.000001.

Value

The same data frame, without features with 0 variance, if that data frame passes all checks.

Details

This function checks that the data frame is not a matrix, that the data frame has more columns than rows (tidy genomic data), that the data frame contains no missing or character values, that no features of the data frame have variance less than epsilon (and removes such features if removeNear0 = TRUE), and checks the data frame for valid column names.

Examples

 # DO NOT CALL THIS FUNCTION DIRECTLY. CALL FROM WITHIN CreateOmics().

 data("colonSurv_df")
 CheckAssay(colonSurv_df[, -(1:3)])

#> # A tibble: 250 x 656
#>      JUN  SOS2  PAK3  RAF1 PRKCB   BTC  SHC1 PRKCA  ELK1  NRG1  PAK2  MTOR  PAK4
#>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1  9.29  5.48  8.21  8.03  5.49  6.65  8.26  8.94  7.38  7.50  7.32  6.96 10.0 
#>  2  9.13  6.35  8.33  7.94  6.26  7.02  8.39  9.61  7.53  7.68  6.80  6.96 10.2 
#>  3  9.37  5.67  7.82  7.74  6.05  7.52  8.69  8.40  7.25  7.33  7.48  7.15  9.48
#>  4 10.6   4.94  8.79  7.64  5.37  6.87  7.81  9.80  7.79  8.38  6.16  6.48 10.6 
#>  5  8.70  5.60  8.75  8.05  6.07  6.49  8.45  8.21  7.60  6.65  7.04  6.66  9.89
#>  6  9.78  5.36  7.56  8.07  5.90  6.39  8.87  8.22  7.35  7.83  7.39  6.90  9.78
#>  7  9.22  5.05  8.20  7.80  5.55  6.86  8.28  8.97  7.43  7.20  7.04  6.96 10.0 
#>  8 10.3   5.33  7.82  7.89  6.27  6.25  8.66  9.71  7.38  7.09  7.22  7.11  9.95
#>  9 10.8   5.07  7.63  7.69  5.48  7.57  8.36  9.69  6.66  7.22  6.99  6.89  9.84
#> 10  9.52  5.50  7.48  7.53  5.71  7.33  8.54  8.14  6.88  7.31  7.01  6.82  9.90
#> # ... with 240 more rows, and 643 more variables: MAP2K4 <dbl>, EIF4EBP1 <dbl>,
#> #   BAD <dbl>, PRKCG <dbl>, NRG3 <dbl>, MAPK9 <dbl>, ERBB4 <dbl>, MAPK10 <dbl>,
#> #   PTK2 <dbl>, ERBB2 <dbl>, ERBB3 <dbl>, MAP2K2 <dbl>, TGFA <dbl>, BRAF <dbl>,
#> #   MAP2K1 <dbl>, MAP2K7 <dbl>, ABL1 <dbl>, NRG2 <dbl>, AKT1 <dbl>, ABL2 <dbl>,
#> #   AKT2 <dbl>, SHC4 <dbl>, RPS6KB1 <dbl>, RPS6KB2 <dbl>, AKT3 <dbl>,
#> #   NRAS <dbl>, GRB2 <dbl>, AREG <dbl>, STAT5B <dbl>, MAPK3 <dbl>,
#> #   STAT5A <dbl>, PAK6 <dbl>, SOS1 <dbl>, MYC <dbl>, MAPK1 <dbl>, NCK1 <dbl>,
#> #   PIK3R5 <dbl>, NRG4 <dbl>, HRAS <dbl>, MAPK8 <dbl>, EGFR <dbl>, GSK3B <dbl>,
#> #   CBLB <dbl>, KRAS <dbl>, CBL <dbl>, SHC3 <dbl>, CDKN1B <dbl>, CDKN1A <dbl>,
#> #   EGF <dbl>, EREG <dbl>, ARAF <dbl>, NCK2 <dbl>, SRC <dbl>, PIK3R3 <dbl>,
#> #   CAMK2A <dbl>, CAMK2B <dbl>, CAMK2D <dbl>, CAMK2G <dbl>, PAK1 <dbl>,
#> #   CBLC <dbl>, CRK <dbl>, PIK3CA <dbl>, PIK3CB <dbl>, CRKL <dbl>,
#> #   PIK3CD <dbl>, GAB1 <dbl>, PLCG1 <dbl>, PLCG2 <dbl>, SHC2 <dbl>,
#> #   HBEGF <dbl>, PIK3CG <dbl>, PIK3R1 <dbl>, PIK3R2 <dbl>, EPHB2 <dbl>,
#> #   EPHB4 <dbl>, EFNA5 <dbl>, PXN <dbl>, CDC42 <dbl>, EFNB3 <dbl>, RRAS <dbl>,
#> #   GRB7 <dbl>, SYNJ1 <dbl>, EPHB3 <dbl>, EFNB1 <dbl>, DNM1 <dbl>,
#> #   MAP4K4 <dbl>, GRIA1 <dbl>, EPHB1 <dbl>, ROCK1 <dbl>, ITSN1 <dbl>,
#> #   RAP1A <dbl>, RAC1 <dbl>, RAP1B <dbl>, EFNB2 <dbl>, WASL <dbl>, TF <dbl>,
#> #   KALRN <dbl>, RASA1 <dbl>, CASP9 <dbl>, E2F1 <dbl>, ...