Manually Create a pathwayCollection-class Object.

Manually create a pathwayCollection list similar to the output of the read_gmt function.

CreatePathwayCollection(pathways, TERMS, ...)

Arguments

pathways
A named list of character vectors. Each vector should contain the names of the individual genes or proteins within that pathway as a vector of character strings. The names contained in these vectors should have non-empty overlap with the feature names of the assay data frame that will be paired with this list in the subsequent analysis. The names of the pathways (the list elements themselves) should be the a shorthand representation of the full pathway name, but this is not required.

TERMS
A character vector the same length as the pathways list with the proper names of the pathways.

...
Additional vectors or data components related to the pathways list. These values should be passed as a name-value pair. See "Details" for more information.

Value

A list object with class pathwayCollection.

Details

This function checks the pathwy list and pathway term inputs and then creates a pathwayCollection object from them. Pass additional list elements (such as the description of each pathway) using the form tag = value through the ... argument (as in the list function). Because some functions in the pathwayPCA package add and edit elements of pathwayCollection objects, please do not create pathwayCollection list items named setsize or n_tested.

See also

Examples

  data("colon_pathwayCollection")

  CreatePathwayCollection(
    pathways = colon_pathwayCollection$pathways,
    TERMS = colon_pathwayCollection$TERMS
  )

#> Object with Class(es) 'pathwayCollection', 'list' [package 'pathwayPCA'] with 2 elements: 
#>  $ pathways:List of 15
#>  $ TERMS   : chr [1:15] "KEGG_PENTOSE_PHOSPHATE_PATHWAY" ...