Merge Phenotype and Assay Data by First Column (Sample ID)

Match the records from the phenotype data to the values in the assay data by sample ID. Return rows from each data frame with matches in both data frames. The sample ID must be the first column in both data frames.

JoinPhenoAssay(pheno_df, assay_df)

Arguments

pheno_df
Phenotype data frame with the sample IDs in the first column

assay_df
Assay data frame with the sample IDs in the first column

Value

A list of three elements:

assay : A data frame with the rows from assay_df which are contained in pheno_df, ordered by their position in pheno_df.

response : A data frame with the rows from pheno_df which are contained in assay_df.

sampleID : A vector of the sample IDs shared by both data frames, ordered by their position in pheno_df.

Details

Don't use this function. This is simply a wrapper around the merge function with extra checks for the class of the ID column. If you want to merge your two data frames by sample ID, you should use the inner_join function from the dplyr package instead. It's easier. See https://dplyr.tidyverse.org/reference/join.html.

Examples

  # DO NOT CALL THIS FUNCTIONS DIRECTLY. USE CreateOmics() INSTEAD.

 data("colonSurv_df")
 JoinPhenoAssay(
   pheno_df = colonSurv_df[, 1:3],
   assay_df = colonSurv_df[, -(2:3)]
 )

#> $assay
#> # A tibble: 250 x 656
#>      JUN  SOS2  PAK3  RAF1 PRKCB   BTC  SHC1 PRKCA  ELK1  NRG1  PAK2  MTOR  PAK4
#>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1  9.29  5.48  8.21  8.03  5.49  6.65  8.26  8.94  7.38  7.50  7.32  6.96 10.0 
#>  2  9.13  6.35  8.33  7.94  6.26  7.02  8.39  9.61  7.53  7.68  6.80  6.96 10.2 
#>  3  9.37  5.67  7.82  7.74  6.05  7.52  8.69  8.40  7.25  7.33  7.48  7.15  9.48
#>  4 10.6   4.94  8.79  7.64  5.37  6.87  7.81  9.80  7.79  8.38  6.16  6.48 10.6 
#>  5  8.70  5.60  8.75  8.05  6.07  6.49  8.45  8.21  7.60  6.65  7.04  6.66  9.89
#>  6  9.78  5.36  7.56  8.07  5.90  6.39  8.87  8.22  7.35  7.83  7.39  6.90  9.78
#>  7  9.22  5.05  8.20  7.80  5.55  6.86  8.28  8.97  7.43  7.20  7.04  6.96 10.0 
#>  8 10.3   5.33  7.82  7.89  6.27  6.25  8.66  9.71  7.38  7.09  7.22  7.11  9.95
#>  9 10.8   5.07  7.63  7.69  5.48  7.57  8.36  9.69  6.66  7.22  6.99  6.89  9.84
#> 10  9.52  5.50  7.48  7.53  5.71  7.33  8.54  8.14  6.88  7.31  7.01  6.82  9.90
#> # ... with 240 more rows, and 643 more variables: MAP2K4 <dbl>, EIF4EBP1 <dbl>,
#> #   BAD <dbl>, PRKCG <dbl>, NRG3 <dbl>, MAPK9 <dbl>, ERBB4 <dbl>, MAPK10 <dbl>,
#> #   PTK2 <dbl>, ERBB2 <dbl>, ERBB3 <dbl>, MAP2K2 <dbl>, TGFA <dbl>, BRAF <dbl>,
#> #   MAP2K1 <dbl>, MAP2K7 <dbl>, ABL1 <dbl>, NRG2 <dbl>, AKT1 <dbl>, ABL2 <dbl>,
#> #   AKT2 <dbl>, SHC4 <dbl>, RPS6KB1 <dbl>, RPS6KB2 <dbl>, AKT3 <dbl>,
#> #   NRAS <dbl>, GRB2 <dbl>, AREG <dbl>, STAT5B <dbl>, MAPK3 <dbl>,
#> #   STAT5A <dbl>, PAK6 <dbl>, SOS1 <dbl>, MYC <dbl>, MAPK1 <dbl>, NCK1 <dbl>,
#> #   PIK3R5 <dbl>, NRG4 <dbl>, HRAS <dbl>, MAPK8 <dbl>, EGFR <dbl>, GSK3B <dbl>,
#> #   CBLB <dbl>, KRAS <dbl>, CBL <dbl>, SHC3 <dbl>, CDKN1B <dbl>, CDKN1A <dbl>,
#> #   EGF <dbl>, EREG <dbl>, ARAF <dbl>, NCK2 <dbl>, SRC <dbl>, PIK3R3 <dbl>,
#> #   CAMK2A <dbl>, CAMK2B <dbl>, CAMK2D <dbl>, CAMK2G <dbl>, PAK1 <dbl>,
#> #   CBLC <dbl>, CRK <dbl>, PIK3CA <dbl>, PIK3CB <dbl>, CRKL <dbl>,
#> #   PIK3CD <dbl>, GAB1 <dbl>, PLCG1 <dbl>, PLCG2 <dbl>, SHC2 <dbl>,
#> #   HBEGF <dbl>, PIK3CG <dbl>, PIK3R1 <dbl>, PIK3R2 <dbl>, EPHB2 <dbl>,
#> #   EPHB4 <dbl>, EFNA5 <dbl>, PXN <dbl>, CDC42 <dbl>, EFNB3 <dbl>, RRAS <dbl>,
#> #   GRB7 <dbl>, SYNJ1 <dbl>, EPHB3 <dbl>, EFNB1 <dbl>, DNM1 <dbl>,
#> #   MAP4K4 <dbl>, GRIA1 <dbl>, EPHB1 <dbl>, ROCK1 <dbl>, ITSN1 <dbl>,
#> #   RAP1A <dbl>, RAC1 <dbl>, RAP1B <dbl>, EFNB2 <dbl>, WASL <dbl>, TF <dbl>,
#> #   KALRN <dbl>, RASA1 <dbl>, CASP9 <dbl>, E2F1 <dbl>, ...
#> 
#> $response
#> # A tibble: 250 x 2
#>    OS_time OS_event
#>      <dbl>    <int>
#>  1  64.9          0
#>  2  59.8          0
#>  3  62.4          0
#>  4  54.5          0
#>  5  46.3          1
#>  6  55.9          0
#>  7  58.0          0
#>  8  54.0          0
#>  9   0.427        1
#> 10  41.4          0
#> # ... with 240 more rows
#> 
#> $sampleID
#>   [1] "subj1"   "subj2"   "subj3"   "subj4"   "subj5"   "subj6"   "subj7"  
#>   [8] "subj8"   "subj9"   "subj10"  "subj11"  "subj12"  "subj13"  "subj14" 
#>  [15] "subj15"  "subj16"  "subj17"  "subj18"  "subj19"  "subj20"  "subj21" 
#>  [22] "subj22"  "subj23"  "subj24"  "subj25"  "subj26"  "subj27"  "subj28" 
#>  [29] "subj29"  "subj30"  "subj31"  "subj32"  "subj33"  "subj34"  "subj35" 
#>  [36] "subj36"  "subj37"  "subj38"  "subj39"  "subj40"  "subj41"  "subj42" 
#>  [43] "subj43"  "subj44"  "subj45"  "subj46"  "subj47"  "subj48"  "subj49" 
#>  [50] "subj50"  "subj51"  "subj52"  "subj53"  "subj54"  "subj55"  "subj56" 
#>  [57] "subj57"  "subj58"  "subj59"  "subj60"  "subj61"  "subj62"  "subj63" 
#>  [64] "subj64"  "subj65"  "subj66"  "subj67"  "subj68"  "subj69"  "subj70" 
#>  [71] "subj71"  "subj72"  "subj73"  "subj74"  "subj75"  "subj76"  "subj77" 
#>  [78] "subj78"  "subj79"  "subj80"  "subj81"  "subj82"  "subj83"  "subj84" 
#>  [85] "subj85"  "subj86"  "subj87"  "subj88"  "subj89"  "subj90"  "subj91" 
#>  [92] "subj92"  "subj93"  "subj94"  "subj95"  "subj96"  "subj97"  "subj98" 
#>  [99] "subj99"  "subj100" "subj101" "subj102" "subj103" "subj104" "subj105"
#> [106] "subj106" "subj107" "subj108" "subj109" "subj110" "subj111" "subj112"
#> [113] "subj113" "subj114" "subj115" "subj116" "subj117" "subj118" "subj119"
#> [120] "subj120" "subj121" "subj122" "subj123" "subj124" "subj125" "subj126"
#> [127] "subj127" "subj128" "subj129" "subj130" "subj131" "subj132" "subj133"
#> [134] "subj134" "subj135" "subj136" "subj137" "subj138" "subj139" "subj140"
#> [141] "subj141" "subj142" "subj143" "subj144" "subj145" "subj146" "subj147"
#> [148] "subj148" "subj149" "subj150" "subj151" "subj152" "subj153" "subj154"
#> [155] "subj155" "subj156" "subj157" "subj158" "subj159" "subj160" "subj161"
#> [162] "subj162" "subj163" "subj164" "subj165" "subj166" "subj167" "subj168"
#> [169] "subj169" "subj170" "subj171" "subj172" "subj173" "subj174" "subj175"
#> [176] "subj176" "subj177" "subj178" "subj179" "subj180" "subj181" "subj182"
#> [183] "subj183" "subj184" "subj185" "subj186" "subj187" "subj188" "subj189"
#> [190] "subj190" "subj191" "subj192" "subj193" "subj194" "subj195" "subj196"
#> [197] "subj197" "subj198" "subj199" "subj200" "subj201" "subj202" "subj203"
#> [204] "subj204" "subj205" "subj206" "subj207" "subj208" "subj209" "subj210"
#> [211] "subj211" "subj212" "subj213" "subj214" "subj215" "subj216" "subj217"
#> [218] "subj218" "subj219" "subj220" "subj221" "subj222" "subj223" "subj224"
#> [225] "subj225" "subj226" "subj227" "subj228" "subj229" "subj230" "subj231"
#> [232] "subj232" "subj233" "subj234" "subj235" "subj236" "subj237" "subj238"
#> [239] "subj239" "subj240" "subj241" "subj242" "subj243" "subj244" "subj245"
#> [246] "subj246" "subj247" "subj248" "subj249" "subj250"
#>