An S4 class for continuous responses within an OmicsPathway object

This creates the OmicsReg class which extends the OmicsPathway master class.

Slots

assayData_df

An \(N \times p\) data frame with named columns.

pathwaySet

A list of known gene pathways with two elements:

pathways : A named list of character vectors. Each vector contains the names of the individual genes within that pathway as a vector of character strings. The names contained in these vectors must have non-empty overlap with the column names of the assayData_df data frame. The names of the pathways (the list elements themselves) should be the a shorthand representation of the full pathway name.

TERMS : A character vector the same length as the pathways list with the proper names of the pathways.

setsize : A named integer vector the same length as the pathways list with the number of genes in each pathway. This list item is calculated during the creation step of a create_OmicsReg function call.

response

A numeric vector of length \(N\): the dependent variable in a regression exercise.

See also