Tabulate, adjust, and sort pathway \(p\)-values

Adjust the pathway \(p\)-values, then return a data frame of the relevant pathway information, sorted by adjusted significance.

TabulatepValues(pVals_vec, genesets_ls, adjust = TRUE,
  proc_vec = c("BH", "Bonferroni", "Holm", "Hochberg", "SidakSS",
  "SidakSD", "BY", "ABH", "TSBH"), ...)

Arguments

pVals_vec
A named vector of permutation \(p\)-values returned by the PermTestSurv, PermTestReg, or PermTestCateg functions when the analysis performed was AES-PCA. Otherwise, when the analysis was performed with Supervised PCA, a named vector of \(p\)-values from the GumbelMixpValues function.

genesets_ls
A list of known gene pathways, trimmed to match the given assay data by the IntersectOmicsPwyCollct function. This pathway list must contain:

pathways : A named list of character vectors where each vector contains the names of the genes in that specific pathway.

TERMS : A character vector the same length as pathways containing the full pathway descriptions.

n_tested : An integer vector the same length as pathways containing the number of genes present in the pathway after trimming. Pathways list trimming is done in the IntersectOmicsPwyCollct function.

adjust
Should you adjust the \(p\)-values for multiple comparisons? Defaults to TRUE.

proc_vec
Character vector of procedures. The returned data frame will be sorted in ascending order by the first procedure in this vector, with ties broken by the unadjusted \(p\)-value. If only one procedure is selected, then it is necessarily the first procedure. Defaults to "BH" (Benjamini and Hochberg, 1995).

...
Additional arguments to pass to the ControlFDR function.

Value

A data frame with columns

pathways : The names of the pathways in the Omics* object (stored in object@trimPathwayCollection$pathways).

n_tested : The number of genes in each pathway after being trimmed to match the assay. Given in the n_tested element of the trimmed pathway collection.

terms : The pathway title, as stored in the object@trimPathwayCollection$TERMS object.

description : The pathway description, if it is stored in the object@trimPathwayCollection$description object.

rawp : The unadjusted \(p\)-values of each pathway.

... : Additional columns as specified through the adjustment argument.

The data frame will be sorted in ascending order by the method specified first in the adjustment argument. If adjustpValues = FALSE, then the data frame will be sorted by the raw \(p\)-values. If you have the suggested tidyverse package suite loaded, then this data frame will print as a tibble. Otherwise, it will stay a simple data frame.

Details

This is a wrapper function for the ControlFDR function. The number of \(p\)-values passed to the pVals_vec argument must equal the number of pathways and set size values in the genesets_ls argument. If you trimmed a pathway from \(p\)- value calculation, then pad this missing value with an NA.

Examples

  # DO NOT CALL THIS FUNCTION DIRECTLY.
  # Call this function through AESPCA_pVals() or SuperPCA_pVals() instead.

  ###  Load the Example Data  ###
  data("colonSurv_df")
  data("colon_pathwayCollection")

  ###  Create an OmicsSurv Object  ###
  colon_Omics <- CreateOmics(
    assayData_df = colonSurv_df[, -(2:3)],
    pathwayCollection_ls = colon_pathwayCollection,
    response = colonSurv_df[, 1:3],
    respType = "surv"
  )

#> 
#>   ======  Creating object of class OmicsSurv  =======

#> The input pathway database included 676 unique features.

#> The input assay dataset included 656 features.

#> Only pathways with at least 3 or more features included in the assay dataset are
#>   tested (specified by minPathSize parameter). There are 15 pathways which meet
#>   this criterion.

#> Because pathwayPCA is a self-contained test (PMID: 17303618), only features in
#>   both assay data and pathway database are considered for analysis. There are 615
#>   such features shared by the input assay and pathway database.

  ###  Extract Pathway PCs and Loadings  ###
  colonPCs_ls <- ExtractAESPCs(
    object = colon_Omics,
    parallel = TRUE,
    numCores = 2
  )

#> Initializing Computing Cluster: 

#> DONE

#> Extracting Pathway PCs in Parallel: 

#> DONE

  ###  Pathway p-Values  ###
  pVals <- PermTestSurv(
    OmicsSurv = colon_Omics,
    pathwayPCs_ls = colonPCs_ls$PCs,
    parallel = TRUE,
    numCores = 2
  )

#> Initializing Computing Cluster: 

#> DONE

#> Extracting Pathway p-Values in Parallel: 

#> DONE

  ###  Create Table of p-Values  ###
  trimmed_PC <- getTrimPathwayCollection(colon_Omics)
  TabulatepValues(
    pVals_vec = pVals,
    genesets_ls = trimmed_PC
  )

#> # A tibble: 15 x 13
#>    pathways n_tested terms    rawp  FDR_BH FWER_Bonferroni FWER_Holm
#>  * <chr>       <int> <chr>   <dbl>   <dbl>           <dbl>     <dbl>
#>  1 pathway~       83 REAC~ 3.38e-4 0.00454         0.00507   0.00507
#>  2 pathway~       40 PID_~ 6.05e-4 0.00454         0.00908   0.00848
#>  3 pathway~      180 REAC~ 1.40e-3 0.00617         0.0211    0.0182 
#>  4 pathway~      104 REAC~ 1.64e-3 0.00617         0.0247    0.0197 
#>  5 pathway~       44 PID_~ 4.13e-3 0.0124          0.0619    0.0454 
#>  6 pathway~       45 KEGG~ 1.21e-2 0.0303          0.182     0.121  
#>  7 pathway~       86 KEGG~ 3.59e-2 0.0770          0.539     0.323  
#>  8 pathway3       26 KEGG~ 1.69e-1 0.317           1         1      
#>  9 pathway~       16 BIOC~ 5.05e-1 0.733           1         1      
#> 10 pathway~       73 KEGG~ 5.08e-1 0.733           1         1      
#> 11 pathway~       23 ST_G~ 5.63e-1 0.733           1         1      
#> 12 pathway~       29 BIOC~ 6.06e-1 0.733           1         1      
#> 13 pathway~       11 BIOC~ 6.36e-1 0.733           1         1      
#> 14 pathway~       54 KEGG~ 8.83e-1 0.933           1         1      
#> 15 pathway~       26 KEGG~ 9.33e-1 0.933           1         1      
#> # ... with 6 more variables: FWER_Hochberg <dbl>, FWER_SidakSS <dbl>,
#> #   FWER_SidakSD <dbl>, FDR_BY <dbl>, FDR_ABH <dbl>, FDR_TSBH_0.05 <dbl>