Colon Cancer -Omics Data
Source: R/data_colonSubset.R

Subset of a colon cancer survival data set, with subject response and assay values.

colonSurv_df

Format

A subset of a data frame containing 656 of 2022 genes measured on 250 subjects. The first two columns are the Overall Survival time (OS_time) and death indicator (OS_event).

Source

Xi Steven Chen