Gene Pathway Subset
Source: R/data_genesetSubset.R

An example Canonical Pathways Gene Subset from the Broad Institute: File: c2.cp.v6.0.symbols.gmt.

colon_pathwaySet

Format

A pathwaySet list of two elements:

pathways : A list of 15 character vectors. Each vector contains the names of the individual genes within that pathway as a vector of character strings.

TERMS : A character vector of length 15 containing the names of the gene pathways.

Source

http://software.broadinstitute.org/gsea/msigdb/collections.jsp

Details

This is a subset of 15 pathways from the Broad Institute gene set list. This subset contains seven pathways which are related to the response information in the colonSurv_df data file.