Generation functions for -Omics*-class objects
Source: R/create_Omics_All.R

These functions create valid objects of class OmicsPathway, OmicsSurv, OmicsReg, or OmicsCateg.

create_OmicsPath(assayData_df, pathwaySet_ls)

create_OmicsSurv(assayData_df, pathwaySet_ls, eventTime_num, eventObserved_lgl)

create_OmicsReg(assayData_df, pathwaySet_ls, response_num)

create_OmicsCateg(assayData_df, pathwaySet_ls, response_fact)

Arguments

assayData_df
An \(N \times p\) data frame with named columns.

pathwaySet_ls
A pathwaySet list of known gene pathways with two elements:

pathways : A named list of character vectors. Each vector contains the names of the individual genes within that pathway as a vector of character strings. The names contained in these vectors must have non-empty overlap with the column names of the assayData_df data frame. The names of the pathways (the list elements themselves) should be the a shorthand representation of the full pathway name.

TERMS: A character vector the same length as the pathways list with the proper names of the pathways.

eventTime_num
A numeric vector with \(N\) observations corresponding to the last observed time of follow up.

eventObserved_lgl
A logical vector with \(N\) observations indicating right-censoring. The values will be FALSE if the observation was censored (i.e., we did not observe an event).

response_num
A numeric vector of length \(N\): the dependent variable in an ordinary regression exercise.

response_fact
A factor vector of length \(N\): the dependent variable of a generalized linear regression exercise.

Value

A valid object of class OmicsPathway, OmicsSurv, OmicsReg, or OmicsCateg.

Details

Please note that the classes of the parameters are not flexible. The -Omics measurement data must be or extend the class data.frame, and the response values (for a survival, regression, or classification object) must match their expected classes exactly. The reason for this is to encourage the end user to pay attention to the quality and format of their input data. Because the functions internal to this package have only been tested on the classes described in the Arguments section, these class checks prevent unexpected errors (or worse, incorrect computational results without an error). These draconian input class restrictions protect the accuracy of your data analysis.

Also note the following: if the supplied pathways object within your pathwaySet_ls list has no names, then this pathway list will be named path1, path2, path3, ...; if any of the pathways are missing names, then the missing pathways will be named noName followed by the index of the pathway. For example, if the 112th pathway in the pathways list has no name (but other pathways do), then this pathway will be named noName112. Furthermore, if any of the pathway names are duplicated, then the duplicates will have .1, .2, .3, ... appended to the duplicate names until all pathway names are unique. Once all pathways have been verified to have unique names, then the pathway names are attached as attributes to the TERMS and setsize vectors (the setsize vector is calculated at object creation).

If your gene pathways list is stored in a .gmt file, use the read_gmt function to import your pathways list as a pathwaySet list object.

OmicsPathway

Valid OmicsPathway objects will have no response information, just the mass spectrometry or bio-assay ("design") matrix and the pathway list. OmicsPathway objects should be created only when unsupervised pathway extraction is needed (not possible with Supervised PCA). Because of the missing response, no pathway testing can be performed on an OmicsPathway object.

OmicsSurv

Valid OmicsSurv objects will have two response vectors: a vector of the most recently recorded follow-up times and a logical vector if that time marks an event (TRUE: observed event; FALSE: right- censored observation).

OmicsReg and OmicsCateg

Valid OmicsReg and OmicsCateg objects with have one response vector of continuous (numeric) or categorial (factor) observations, respectively.

See also

Examples

# NOT RUN {
  ###  Load the Example Data  ###
  data("colonSurv_df")
  data("colon_pathwaySet")

  ###  Create an OmicsPathway Object  ###
  colon_OmicsPath <- create_OmicsPath(assayData_df = colonSurv_df[, -(1:2)],
                                      pathwaySet_ls = colon_pathwaySet)

  ###  Create an OmicsSurv Object  ###
  colon_OmicsSurv <- create_OmicsSurv(assayData_df = colonSurv_df[, -(1:2)],
                                      pathwaySet_ls = colon_pathwaySet,
                                      eventTime_num = colonSurv_df$OS_time,
                                      eventObserved_lgl = as.logical(colonSurv_df$OS_event))

  ###  Create an OmicsReg Object  ###
  colon_OmicsReg <- create_OmicsReg(assayData_df = colonSurv_df[, -(1:2)],
                                    pathwaySet_ls = colon_pathwaySet,
                                    response_num = colonSurv_df$OS_time)

  ###  Create an OmicsCateg Object  ###
  colon_OmicsCateg <- create_OmicsCateg(assayData_df = colonSurv_df[, -(1:2)],
                                        pathwaySet_ls = colon_pathwaySet,
                                        response_fact = as.factor(colonSurv_df$OS_event))
# }