Manually Create a pathwaySet-class Object.
Source: R/create_pathwaySet.R

Manually create a pathwaySet list similar to the output of the read_gmt function.

create_pathwaySet(pathways, TERMS, ...)

Arguments

pathways
A named list of character vectors. Each vector should contain the names of the individual genes or proteins within that pathway as a vector of character strings. The names contained in these vectors should have non-empty overlap with the feature names of the assay data frame that will be paired with this list in the subsequent analysis. The names of the pathways (the list elements themselves) should be the a shorthand representation of the full pathway name, but this is not required.

TERMS
A character vector the same length as the pathways list with the proper names of the pathways.

...
Additional vectors or data components related to the pathways list. These values should be passed as a name-value pair. See "Details" for more information.

Value

A list object with class pathwaySet.

Details

This function checks the pathwy list and pathway term inputs and then creates a pathwaySet object from them. Pass additional list elements using the form tag = value through the ... argument (as in the list function). Because some functions in the pathwayPCA package add and edit elements of pathwaySet objects, please do not create pathwaySet list items named setsize or trim_setsize.

See also

Examples

  data("colon_pathwaySet")

  create_pathwaySet(pathways = colon_pathwaySet$pathways,
                    TERMS = colon_pathwaySet$TERMS)

#> Object with Class(es) 'pathwaySet', 'list' [package 'pathwayPCA'] with 2 elements: 
#>  $ pathways:List of 15
#>  $ TERMS   : chr [1:15] "KEGG_PENTOSE_PHOSPHATE_PATHWAY" ...