Extract AES-PCs from expressed pathway-subsets of a mass spectrometry or bio-assay data frame

Given a clean OmicsPath object (cleaned by the expressedOmes function), extract the first principal components from each expressed pathway in the assay design matrix.

extract_aesPCs(object, trim = 3, numPCs = 1, parallel = FALSE,
  numCores = NULL, ...)

# S4 method for OmicsPathway
extract_aesPCs(object, trim = 3, numPCs = 1,
  parallel = FALSE, numCores = NULL, ...)

Arguments

object
An object of class OmicsPathway.

trim
The minimum cutoff of expressed -Ome measures before a pathway is excluded. Defaults to 3.

numPCs
The number of PCs to extract from each pathway. Defaults to 1.

parallel
Should the comuptation be completed in parallel? Defaults to FALSE.

numCores
If parallel = TRUE, how many cores should be used for computation? Defaults to NULL.

...
Dots for additional internal arguments (currently unused).

Value

A list of matrices. Each element of the list will be named by its pathway, and the elements will be \(N \times\) numPCs matrices containing the first numPCs principal components from each pathway. See "Details" for more information.

Details

This function takes in a data frame with named columns and a pathway list as an OmicsPathway object which has had unexpressed -Omes removed by the expressedOmes function. This function will then iterate over the list of pathways, extracting columns from the assay design matrix which match the genes listed in that pathway as a sub-matrix (as a data.frame object). This function will then call the aespca on each data frame in the list of pathway-specific design matrices, extracting the first numPCs AES principal components from each pathway data frame. These PC matrices are returned as a named list.

NOTE: some genes will be included in more than one pathway, so these pathways are not mutually exclusive. Further note that there may be many genes in the assay design matrix that are not included in the pathway sets, so these will not be extracted to the list. It is then vitally important to use either a very broad and generic pathway set list or a pathway set list that is appropriate for the assay data supplied.

See also

create_OmicsPath; expressedOmes; aespca

Examples

  # DO NOT CALL THIS FUNCTION DIRECTLY.
  # Use AESPCA_pVals() instead