Extract Table of \(p\)-values from a superpcOut- or aespcOut- class Object.

Given an object of class aespcOut or superpcOut, as returned by the functions AESPCA_pVals or SuperPCA_pVals, respectively, return a data frame of the \(p\)-values for the top pathways.

getPathpVals(pcOut, score = FALSE, numPaths = 20L, alpha = NULL, ...)

# S3 method for superpcOut
getPathpVals(pcOut, score = FALSE, numPaths = 20L,
  alpha = NULL, ...)

# S3 method for aespcOut
getPathpVals(pcOut, score = FALSE, numPaths = 20L,
  alpha = NULL, ...)

Arguments

pcOut
An object of classes superpcOut or aespcOut as returned by the SuperPCA_pVals or AESPCA_pVals functions, respectively.

score
Should the unadjusted \(p\)-values be returned transformed to negative natural logarithm scores or left as is? Defaults to FALSE; that is, the raw \(p\)-values are returned instead of the transformed \(p\)-values.

numPaths
The number of top pathways by raw \(p\)-value. Defaults to the top 20 pathways. We do not permit users to specify numPaths and alpha concurrently.

alpha
The significance threshold for raw \(p\)-values. Defaults to NULL. If alpha is given, then numPaths will be ignored.

...
Dots for additional arguments (currently unused).

Value

A data frame with the following columns:

terms : The pathway name, as given in the object@trimPathwayCollection$TERMS object.

description : (OPTIONAL) The pathway description, as given in the object@trimPathwayCollection$description object, if supplied.

rawp : The unadjusted \(p\)-values of each pathway. Included if score = FALSE.

... : Additional columns of FDR-adjusted \(p\)-values as specified through the adjustment argument of the SuperPCA_pVals or AESPCA_pVals functions.

score : The negative natural logarithm of the unadjusted \(p\)-values of each pathway. Included if score = TRUE.

NULL

NULL

Details

Row-subset the pVals_df entry of an object of class aespcOut or superpcOut by the number of pathways requested (via the nPaths argument) or by the unadjusted significance level for each pathway (via the alpha argument). Return a data frame of the pathway names, FDR-adjusted significance levels (if available), and the raw score (negative natural logarithm of the \(p\)-values) of each pathway.

Examples

  ###  Load Data  ###
  data("colonSurv_df")
  data("colon_pathwayCollection")

  ###  Create -Omics Container  ###
  colon_Omics <- CreateOmics(
    assayData_df = colonSurv_df[, -(2:3)],
    pathwayCollection_ls = colon_pathwayCollection,
    response = colonSurv_df[, 1:3],
    respType = "survival"
  )

#> 
#>   ======  Creating object of class OmicsSurv  =======

#> The input pathway database included 676 unique features.

#> The input assay dataset included 656 features.

#> Only pathways with at least 3 or more features included in the assay dataset are
#>   tested (specified by minPathSize parameter). There are 15 pathways which meet
#>   this criterion.

#> Because pathwayPCA is a self-contained test (PMID: 17303618), only features in
#>   both assay data and pathway database are considered for analysis. There are 615
#>   such features shared by the input assay and pathway database.

  ###  Calculate Supervised PCA Pathway p-Values  ###
  colon_superpc <- SuperPCA_pVals(
    colon_Omics,
    numPCs = 2,
    parallel = TRUE,
    numCores = 2,
    adjustment = "BH"
  )

#> Initializing Computing Cluster: 

#> DONE

#> Calculating Pathway Test Statistics in Parallel: 

#> DONE

#> Calculating Pathway Critical Values in Parallel: 

#> DONE

#> Calculating Pathway p-Values: 

#> DONE

#> Adjusting p-Values and Sorting Pathway p-Value Data Frame: 

#> DONE

  ###  Extract Table of p-Values  ###
  # Top 5 Pathways
  getPathpVals(
    colon_superpc,
    numPaths = 5
  )

#> # A tibble: 5 x 3
#>   terms                                     rawp FDR_BH
#>   <chr>                                    <dbl>  <dbl>
#> 1 KEGG_ERBB_SIGNALING_PATHWAY            0.00897 0.0396
#> 2 REACTOME_SIGNALING_BY_INSULIN_RECEPTOR 0.0101  0.0396
#> 3 PID_TNF_PATHWAY                        0.0123  0.0396
#> 4 REACTOME_PHOSPHOLIPID_METABOLISM       0.0156  0.0396
#> 5 PID_EPHB_FWD_PATHWAY                   0.0158  0.0396

  # Pathways with Unadjusted p-Values < 0.01
  getPathpVals(
    colon_superpc,
    alpha = 0.01
  )

#> # A tibble: 1 x 3
#>   terms                          rawp FDR_BH
#>   <chr>                         <dbl>  <dbl>
#> 1 KEGG_ERBB_SIGNALING_PATHWAY 0.00897 0.0396