Access and Edit Response of an OmicsReg or OmicsReg Object

"Get" or "Set" the values of the response_num or response_fact slots of an object of class OmicsReg or OmicsReg, respectively.

getResponse(object, ...)

getResponse(object) <- value

# S4 method for OmicsPathway
getResponse(object, ...)

# S4 method for OmicsPathway
getResponse(object) <- value

Arguments

object
An object of class OmicsReg-class or OmicsCateg-class.

...
Dots for additional internal arguments (currently unused).

value
The replacement object to be assigned to the response slot.

Value

The "get" functions return the objects in the slots specified: getResponse returns the response_num vector from objects of class OmicsReg and the response_fact vector from objects of class OmicsCateg. These functions can extract these values from any valid object of those classes.

The "set" functions enable the user to edit or replace the object in the response_num slot for any OmicsReg object or response_fact slot for any OmicsCateg object, provided that the new values do not violate the validity check of such an object. See "Details" for more information.

Details

These functions can be useful to set or extract the response vector from an object of class OmicsReg or OmicsReg. However, we recommend that users simply create a new, valid object instead of modifying an existing one. The validity of edited objects is checked with their respective ValidOmicsCateg or ValidOmicsReg function. Because both classes have a response slot, we set this method for the parent class, OmicsPathway-class.

See also

Examples

  data("colonSurv_df")
  data("colon_pathwayCollection")

  colon_Omics <- CreateOmics(
    assayData_df = colonSurv_df[, -(2:3)],
    pathwayCollection_ls = colon_pathwayCollection,
    response = colonSurv_df[, c(1, 2)],
    respType = "reg"
  )

#> 
#>   ======  Creating object of class OmicsReg  =======

#> The input pathway database included 676 unique features.

#> The input assay dataset included 656 features.

#> Only pathways with at least 3 or more features included in the assay dataset are
#>   tested (specified by minPathSize parameter). There are 15 pathways which meet
#>   this criterion.

#> Because pathwayPCA is a self-contained test (PMID: 17303618), only features in
#>   both assay data and pathway database are considered for analysis. There are 615
#>   such features shared by the input assay and pathway database.

  getResponse(colon_Omics)

#>   [1]  64.8657534  59.7698630  62.4000000  54.5095890  46.2904110  55.8575343
#>   [7]  57.9616438  54.0493151   0.4273973  41.4246575  40.7342466  68.0547945
#>  [13]  21.6328767  33.1068493  68.6136986  49.4465753  64.1753425  58.2575342
#>  [19]  53.2602740  63.9123288  60.0328767  60.0328767  60.5589041  57.0739726
#>  [25]  54.7397260  48.2630137  57.2383562  70.4547945   7.8246575  50.1369863
#>  [31]  46.9808219  46.4547945  40.6684932  13.8410959 111.4849315  52.1095890
#>  [37]  54.1808219  50.2027397   1.5452055  17.6547945  12.6575342   6.9041096
#>  [43]  32.5150685  16.1424657  69.1726027  70.8493151  74.3013699  76.7671233
#>  [49]  23.4082192  36.0000000   5.2273973   8.1534247   3.6164384  30.8383562
#>  [55]  40.9972603 142.5500000 122.7200000  68.5100000 123.5800000  28.9600000
#>  [61] 119.2100000  59.5300000  68.8100000  82.2900000 110.8200000  65.8800000
#>  [67]  84.1300000  85.4400000  64.3700000  16.4700000  89.7500000  66.5700000
#>  [73]  76.0700000  67.8200000  52.2400000  64.1000000  55.1300000  57.7900000
#>  [79]  50.1000000  43.7900000  30.2700000   3.6400000  36.5900000  20.0200000
#>  [85] 118.5800000  55.3300000 134.8600000 127.2900000  13.3800000 120.6200000
#>  [91] 106.9400000  18.2100000  11.7000000 111.6400000  28.7900000  26.8200000
#>  [97]  52.5000000  24.0300000  88.9900000 102.4400000  33.7300000  85.2800000
#> [103]  26.5300000  37.3100000  86.4300000  55.9200000  70.6500000  76.0400000
#> [109]  79.1300000  58.0200000  83.9900000  80.2500000  40.6000000  95.3700000
#> [115]  31.0600000  37.3100000  22.0200000  63.2500000  34.0900000  52.1400000
#> [121]  66.0400000  59.9600000  47.8600000  58.4500000  45.9900000  60.4200000
#> [127]  37.6700000  15.7100000  39.2500000  49.8700000  25.2400000  44.9700000
#> [133]  35.9000000  26.6600000  36.7800000  41.5200000  22.2200000  33.2000000
#> [139]  39.6800000  25.6100000  83.9900000  15.0500000   8.3800000  18.1800000
#> [145]  27.1500000  24.2600000 105.1700000  11.7600000  70.4500000  95.0700000
#> [151] 112.3300000  27.5100000  45.9200000  54.0400000  49.8700000 103.9800000
#> [157]   6.1100000   8.0500000  99.5100000  20.4400000  34.9800000  76.6000000
#> [163]  10.0200000  41.4500000  96.2600000  85.6100000  78.8700000  22.2500000
#> [169]  74.5300000  59.0700000  47.8000000   5.6800000  45.8300000  74.3600000
#> [175]  79.3900000  74.0700000  50.7600000  67.8200000  70.8400000  50.5900000
#> [181]  34.5800000  38.7200000  54.9000000  53.6800000  67.9200000  58.4500000
#> [187]  59.3400000  54.5400000  19.0300000  26.2600000  52.8600000  32.0500000
#> [193]  55.9500000  56.6700000  50.4300000   4.2400000  49.1100000  49.5700000
#> [199]  50.4900000  63.0500000  49.2400000   0.9200000  31.3600000  40.3000000
#> [205] 116.6400000  31.9200000  19.9800000  24.6200000  18.5400000 130.7500000
#> [211]  17.6200000   5.3200000   2.3600000   5.7200000  15.2800000  34.4500000
#> [217]  95.0700000   4.9600000  28.8900000   3.2500000   0.9200000  18.7700000
#> [223]  57.7300000  13.4700000  14.3900000   5.1200000   4.9900000  33.1700000
#> [229]  34.8400000  50.5300000   5.7200000   8.0200000  34.0200000   1.4400000
#> [235]  36.9500000  60.7500000  37.4400000  23.0700000  19.8500000  31.8500000
#> [241]   8.6100000  42.2700000  15.2500000  14.5600000   9.9600000   8.9400000
#> [247]  32.6700000   2.0700000  22.7800000  22.3500000