Positive root of a symmetric matrix
Source: R/calculate_matrixRoot.R

Calculate the matrix root of a symmetric matrix via the Spectral Decomposition.

matrixRoot(x, root = 2)

Arguments

x
A symmetric (necessarily square) matrix.

root
A positive real number.

Value

A matrix, that when multiplied by itself root times, yields x.

Details

This function decomposes x into \(V \times D \times V^T\) via the eigen function, sets any numerically negative eigenvalues to 0, calculates the root of these eigenvalues as \(D^r\), then returns the matrix \(V \times D^r \times V^T\).

See https://en.wikipedia.org/wiki/Eigendecomposition_of_a_matrix.

Examples

  X <- matrix(rnorm(25), ncol = 5);   xTx <- t(X) %*% X
  matrixRoot(xTx)

#>            [,1]       [,2]       [,3]       [,4]       [,5]
#> [1,]  2.6369879 -0.3646567 -0.6520587  0.5001262  0.7680723
#> [2,] -0.3646567  2.0837803 -0.5157137 -0.2651139 -0.1650258
#> [3,] -0.6520587 -0.5157137  2.6158884 -0.2056113 -0.5594172
#> [4,]  0.5001262 -0.2651139 -0.2056113  2.0121197 -0.6820119
#> [5,]  0.7680723 -0.1650258 -0.5594172 -0.6820119  2.1836310

  matrixRoot(xTx, root = 3)

#>            [,1]        [,2]       [,3]       [,4]        [,5]
#> [1,]  1.8599266 -0.18578299 -0.3074680  0.2704974  0.39159869
#> [2,] -0.1857830  1.60927161 -0.2764877 -0.1446830 -0.09676689
#> [3,] -0.3074680 -0.27648770  1.8628072 -0.1156663 -0.27794665
#> [4,]  0.2704974 -0.14468297 -0.1156663  1.5512579 -0.38724886
#> [5,]  0.3915987 -0.09676689 -0.2779466 -0.3872489  1.62270223