Calculate pathway-specific Student's \(t\)-scores for supervised PCA

Extract principal components (PCs) from the gene pathway, and return the test statistics associated with the first numPCs principal components at a set of threshold values.

pathway_tScores(pathway_vec, geneArray_df, response_mat,
  responseType = c("survival", "regression", "categorical"),
  n.threshold = 20, numPCs = 1, min.features = 3)

Arguments

pathway_vec
A character vector of the measured -Omes in the chosen gene pathway. These should match a subset of the rownames of the gene array.

geneArray_df
A "tall" pathway data frame (\(p \times N\)). Each subject or tissue sample is a column, and the rows are the -Ome measurements for that sample.

response_mat
A response matrix corresponding to responseType. For "regression" and "categorical", this will be an \(N \times 1\) matrix of response values. For "survival", this will be an \(N \times 2\) matrix with event times in the first column and observed event indicator in the second.

responseType
A character string. Options are "survival", "regression", and "categorical".

n.threshold
The number of bins into which to split the feature scores in the fit object returned internally by the superpc.train function.

numPCs
The number of PCs to extract from the pathway.

min.features
What is the smallest number of genes allowed in each pathway? This argument must be kept constant across all calls to this function which use the same pathway list. Defaults to 3.

Value

A matrix with numPCs rows and n.threshold columns. The matrix values are model \(t\)-statisics for each PC included (rows) at each threshold level (columns).

Details

This is a wrapper function to call superpc.train and superpc.st. This wrapper is designed to facilitate apply calls (in parallel or serially) of these two functions over a list of gene pathways. When numPCs is equal to 1, we recommend using a simplify-style apply variant, such as sapply (shown in lapply) or parSapply (shown in clusterApply), then transposing the resulting matrix.

See also

Examples

  # DO NOT CALL THIS FUNCTION DIRECTLY.
  # Use SuperPCA_pVals() instead