Test pathways with Supervised PCA

Given a supervised OmicsPath object (one of OmicsSurv, OmicsReg, or OmicsCateg), extract the first \(k\) principal components (PCs) from each pathway-subset of the -Omics assay design matrix, test their association with the response matrix, and return a data frame of the adjusted \(p\)-values for each pathway.

SuperPCA_pVals(object, n.threshold = 20, numPCs = 1,
  parallel = FALSE, numCores = NULL, adjustpValues = TRUE,
  adjustment = c("Bonferroni", "Holm", "Hochberg", "SidakSS", "SidakSD",
  "BH", "BY", "ABH", "TSBH"), ...)

# S4 method for OmicsPathway
SuperPCA_pVals(object, n.threshold = 20,
  numPCs = 1, parallel = FALSE, numCores = NULL,
  adjustpValues = TRUE, adjustment = c("Bonferroni", "Holm",
  "Hochberg", "SidakSS", "SidakSD", "BH", "BY", "ABH", "TSBH"), ...)

Arguments

object
An object of superclass OmicsPathway with a response matrix or vector.

n.threshold
The number of bins into which to split the feature scores in the fit object returned internally by the superpc.train function to the pathway_tScores and pathway_tControl functions. Defaults to 20. Smaller values may result in less accurate pathway \(p\)-values while larger values increase computation time.

numPCs
The number of PCs to extract from each pathway. Defaults to 1.

parallel
Should the computation be completed in parallel? Defaults to FALSE.

numCores
If parallel = TRUE, how many cores should be used for computation? Internally defaults to the number of available cores minus 1.

adjustpValues
Should you adjust the \(p\)-values for multiple comparisons? Defaults to TRUE.

adjustment
Character vector of procedures. The returned data frame will be sorted in ascending order by the first procedure in this vector, with ties broken by the unadjusted \(p\)-value. If only one procedure is selected, then it is necessarily the first procedure. See the documentation for the ControlFDR function for the adjustment procedure definitions and citations.

...
Dots for additional internal arguments.

Value

A data frame with columns:

pathways : The names of the pathways in the Omics* object (given in object@trimPathwayCollection$pathways.)

setsize : The number of genes in each of the original pathways (given in the object@trimPathwayCollection$setsize object).

terms : The pathway description, as given in the object@trimPathwayCollection$TERMS object.

rawp : The unadjusted \(p\)-values of each pathway.

... : Additional columns as specified through the adjustment argument.

The data frame will be sorted in ascending order by the method specified first in the adjustment argument. If adjustpValues = FALSE, then the data frame will be sorted by the raw \(p\)-values. If you have the suggested tidyverse package suite loaded, then this data frame will print as a tibble. Otherwise, it will print as a data frame.

Details

This is a wrapper function for the pathway_tScores, pathway_tControl, OptimGumbelMixParams, GumbelMixpValues, and TabulatepValues functions.

Please see our Quickstart Guide for this package: https://gabrielodom.github.io/pathwayPCA/articles/Supplement1-Quickstart_Guide.html

See also

Examples

  ###  Load the Example Data  ###
  data("colonSurv_df")
  data("colon_pathwayCollection")

  ###  Create an OmicsSurv Object  ###
  colon_OmicsSurv <- CreateOmics(
    assayData_df = colonSurv_df[, -(2:3)],
    pathwayCollection_ls = colon_pathwayCollection,
    response = colonSurv_df[, 1:3],
    respType = "surv"
  )

#> 
#>   ======  Creating object of class OmicsSurv  =======

#> The input pathway database included 676 unique features.

#> The input assay dataset included 656 features.

#> Only pathways with at least 3 or more features included in the assay dataset are
#>   tested (specified by minPathSize parameter). There are 15 pathways which meet
#>   this criterion.

#> Because pathwayPCA is a self-contained test (PMID: 17303618), only features in
#>   both assay data and pathway database are considered for analysis. There are 615
#>   such features shared by the input assay and pathway database.

  ###  Calculate Pathway p-Values  ###
  colonSurv_superpc <- SuperPCA_pVals(
    object = colon_OmicsSurv,
    parallel = TRUE,
    numCores = 2,
    adjustpValues = TRUE,
    adjustment = c("Hoch", "SidakSD")
  )

#> Initializing Computing Cluster: 

#> DONE

#> Calculating Pathway Test Statistics in Parallel: 

#> DONE

#> Calculating Pathway Critical Values in Parallel: 

#> DONE

#> Calculating Pathway p-Values: 

#> DONE

#> Adjusting p-Values and Sorting Pathway p-Value Data Frame: 

#> DONE