Transpose an Assay (Data Frame)
Source: R/transpose_assay.R

Transpose an object of class data.frame that contains assay measurements while preserving row (feature) and column (sample) names.

transpose_assay(assay_df, firstColIsFeatureNames = TRUE)

Arguments

assay_df
A data frame with numeric values to transpose

firstColIsFeatureNames
Are the data row names in the first column of df? Defaults to TRUE. If not, this function assumes that the data row names are accesible by the rownames function.

Value

The transposition of df, with row and column names preserved.

Details

This function is designed to transpose "tall" assay data frames (where genes or proteins are the rows and patient or tumour samples are the columns). This function also transposes the row (feature) names to column names and the column (sample) names to row names. Notice that all rows and columns (other than the feature name column, as applicable) are numeric.

Examples

   x_mat <- matrix(rnorm(5000), ncol = 20, nrow = 250)
   rownames(x_mat) <- paste0("gene_", 1:250)
   colnames(x_mat) <- paste0("sample_", 1:20)
   x_df <- as.data.frame(x_mat, row.names = rownames(x_mat))

   transpose_assay(x_df, firstColIsFeatureNames = FALSE)

#>                gene_1      gene_2      gene_3       gene_4       gene_5
#> sample_1  -0.09744510 -0.93584735 -0.01595031 -0.826788954 -1.512399651
#> sample_2   1.87774387  2.15875655  0.70971452  0.766983379 -0.308211421
#> sample_3  -0.65954237  0.05742171 -2.80801051 -0.912259753 -0.782379163
#> sample_4   0.93312514 -0.96127668  0.25508171 -0.545401573  0.930360736
#>                 gene_6     gene_7      gene_8      gene_9      gene_10
#> sample_1   0.935363190  0.1764886  0.24368546  1.62354888  0.112038083
#> sample_2   1.012001849 -0.9190516  0.56338008  0.32248275  0.366674363
#> sample_3  -0.664104924  0.6263098 -0.50724821  0.27036133  0.467476865
#> sample_4  -0.537650558 -0.4524261 -0.43929097 -0.61622311  0.441634552
#>               gene_11    gene_12    gene_13     gene_14     gene_15     gene_16
#> sample_1  -0.13399701 -1.9100875 -0.2792372 -0.31344598  1.06730788  0.07003485
#> sample_2   1.12983515 -0.9414981  0.2178376  1.41541229 -0.38373305 -0.17408637
#> sample_3   0.72399496  0.6138369 -0.6178692  0.22072490  1.12792660  1.81345434
#> sample_4   0.48259745  0.5421444 -2.2907847  0.31035185  1.40407898  1.37711764
#>               gene_17    gene_18     gene_19    gene_20      gene_21
#> sample_1  -0.63912332 -0.0499649 -0.25148344  0.4447971  2.755417575
#> sample_2  -0.22174452 -1.0095287  0.48072527  1.6044073 -1.515024529
#> sample_3  -0.08382568  1.3677067 -0.62743462 -0.2166291 -0.683713824
#> sample_4   1.06039565  0.6321729  1.08492804  1.3564594  0.362424040
#>               gene_22     gene_23     gene_24     gene_25      gene_26
#> sample_1   0.04653138  0.57770907  0.11819487 -1.91172049  0.862086482
#> sample_2  -1.41602391  0.87677733  0.62413241  2.11227729 -0.356124416
#> sample_3  -0.44470273  0.60648981  0.62418307 -0.69543107 -0.783639078
#> sample_4   2.16934446  0.13913051  1.37632653 -0.49144999  1.534915763
#>               gene_27     gene_28     gene_29     gene_30       gene_31
#> sample_1  -0.24323674 -0.20608719  0.01917759  0.02956075  0.5498275418
#> sample_2  -1.06446421  1.07711654  1.18157557  0.19839209 -0.4004052494
#> sample_3  -0.95312386  1.79275607  0.34897670  0.25910377 -0.8059518969
#> sample_4  -0.41619872 -0.52054380  0.85058387  0.33449657 -0.8293517745
#>              gene_32     gene_33    gene_34     gene_35      gene_36
#> sample_1  -2.2741149  2.68255718 -0.3612213  0.21335575  1.074345882
#> sample_2   0.6161543  1.97415675  1.8846623 -1.58862055 -0.539923164
#> sample_3   0.1056647 -0.33359968  1.6418480 -0.64390586  0.587020562
#> sample_4  -0.2186959 -1.54508372  0.2332298  0.03106964  0.357865645
#>               gene_37    gene_38    gene_39     gene_40     gene_41     gene_42
#> sample_1  -0.66508825  1.1139524 -0.2458964 -1.17756331 -0.97585062  1.06505732
#> sample_2  -1.16946146  0.5591060 -1.8193472  0.39334397  0.04213411  1.17966418
#> sample_3  -0.15040309 -1.7108218  1.4310326 -2.64521227 -1.03245741 -0.70746643
#> sample_4   1.60862422  1.4298543 -0.9483396  1.01570554  0.03773461 -1.74423940
#>               gene_43     gene_44    gene_45      gene_46    gene_47
#> sample_1   0.13167063  0.48862881 -1.6994506 -1.470736306  0.2841503
#> sample_2  -0.25692118 -1.05633610  0.1987772  0.650533552  0.3439133
#> sample_3  -0.70056001  0.53788544 -0.3163322 -0.839622754 -1.3549281
#> sample_4  -0.97007381 -0.22902736 -0.9748457  2.692372420  1.3923939
#>               gene_48     gene_49     gene_50    gene_51     gene_52
#> sample_1   1.33732041  0.23669628  1.31829338  0.5239098  0.60674805
#> sample_2   1.47753231  0.07202570  2.12644453 -1.4761969  0.40788850
#> sample_3  -0.81756827 -0.63440000  0.81594943  0.3027957  1.80708662
#> sample_4   1.36007615 -0.36541588 -0.86893720 -0.5065567  1.21470730
#>               gene_53     gene_54     gene_55    gene_56     gene_57
#> sample_1  -0.10993567  0.17218172 -0.09032729  1.9243433  1.29839276
#> sample_2   1.39397780  0.36027830  0.65455025  1.0521554 -1.97955513
#> sample_3  -0.89402676 -0.04642821 -0.47117914 -0.5266926 -0.09513491
#> sample_4   0.50695868 -2.09497131  0.03407924  0.8527287  0.74322814
#>               gene_58     gene_59     gene_60     gene_61     gene_62
#> sample_1   0.74879127  0.55622433 -0.54825726  1.11053489 -2.61233433
#> sample_2   1.20838561 -0.16928008  0.29502975  1.26634059 -1.13534326
#> sample_3  -2.49536481  0.16688922  0.35049238  1.43370101  0.76590680
#> sample_4   0.55715361 -1.24858322 -0.20787431  0.42396946 -0.50669744
#>              gene_63       gene_64     gene_65    gene_66    gene_67
#> sample_1  -0.1556938  0.4338897899 -0.38195111  0.4241876  1.0631020
#> sample_2  -1.1310538  0.1099933842  0.85290541 -0.2343379  2.0866886
#> sample_3   1.1675207 -0.1369434286 -0.51490204  1.5197445 -0.3284917
#> sample_4  -0.6115233  2.4116594528 -0.16495814 -0.4404060  0.5219762
#>               gene_68     gene_69     gene_70     gene_71     gene_72
#> sample_1   1.04871262 -0.03810289  0.48614892  1.67288261 -0.35436116
#> sample_2  -0.11091937 -1.39284706 -1.14229077  1.70460874 -0.08007363
#> sample_3  -0.05367151 -0.56352463 -0.74390896 -0.10904165 -0.56082923
#> sample_4  -1.91832225 -1.98264996  0.52120016  0.77778320 -0.81211887
#>               gene_73    gene_74     gene_75    gene_76     gene_77    gene_78
#> sample_1   0.94634789  1.3168264 -0.29664002 -0.3872136 -0.78543266 -1.0567369
#> sample_2  -0.43728124 -0.1192151  0.78646286 -0.5789452 -0.14542689  0.5264580
#> sample_3   0.18800155  0.7488509 -1.91653832  0.2360958  0.62895342  0.4179257
#> sample_4   0.91074832  0.9487540 -1.34804945  0.3541759  0.53026393 -0.3109749
#>               gene_79     gene_80    gene_81    gene_82      gene_83
#> sample_1  -0.79554143 -1.75627543 -0.6905379 -0.5585420 -0.536663326
#> sample_2   1.73357811  1.44865722  1.5181931 -0.3840073  1.827125177
#> sample_3   1.97675848 -0.50628630 -1.1099689 -0.9487057  0.476843757
#> sample_4  -0.24415553 -0.29187819 -1.1280862 -1.1228816  2.053938640
#>               gene_84     gene_85     gene_86     gene_87       gene_88
#> sample_1   0.22712713  0.97845492 -0.20888265 -1.39941046  0.2585372879
#> sample_2  -0.55149175 -0.86575354 -0.34383148  1.06287646  0.8130582040
#> sample_3  -0.79520156  0.23432692 -1.22245110 -2.45364735 -1.4892608139
#> sample_4  -0.90951843  0.45837916 -0.04661845 -0.68095460 -1.4888294023
#>               gene_89    gene_90     gene_91     gene_92     gene_93
#> sample_1  -0.44179945  0.5685999  2.12685046  0.42485844 -1.68428153
#> sample_2   1.80348336 -0.1050687  0.98245336 -1.71330262 -0.83201953
#> sample_3  -0.43214773 -0.9425540 -0.12145080  1.33644680 -0.86035618
#> sample_4  -0.39251301 -1.7496820 -0.03866763 -0.79715324 -0.91592054
#>               gene_94    gene_95      gene_96     gene_97     gene_98
#> sample_1   0.24940178  1.0728383  2.039369263  0.44945378  1.39181405
#> sample_2   1.10049188 -0.1738201  0.178812018 -0.69842945 -0.96044916
#> sample_3   0.66653782 -1.4215347  1.170056168 -1.40471454  1.10170810
#> sample_4   0.07326746  1.2381713  0.718371336  0.53946650 -1.06123399
#>               gene_99    gene_100    gene_101    gene_102    gene_103
#> sample_1   0.42656655  0.10758399  0.02229473  0.60361101 -0.26265057
#> sample_2  -0.97542304 -0.33857650  1.15234707  0.40510118 -0.47092250
#> sample_3   0.69798626 -0.86434980 -1.09147035 -0.03705146  0.81005379
#> sample_4  -0.54073787 -0.71520471 -0.17892009  0.16497028 -0.72740817
#>              gene_104   gene_105    gene_106   gene_107    gene_108    gene_109
#> sample_1  -0.52826408  0.1921494 -1.14619967  0.8461847  0.08171963 -1.30511701
#> sample_2  -0.13325102  1.2266824  0.33294400 -0.3470885 -0.09855069  0.03476606
#> sample_3  -0.49935541  0.9480316 -0.17424596 -1.1062360 -0.94598501  0.28908959
#> sample_4  -0.53783713 -0.5996649  1.18208775  0.1892129  0.24894911  1.04664341
#>             gene_110    gene_111     gene_112    gene_113    gene_114
#> sample_1  -0.9449121  0.45434159 -0.855202501 -0.28689522  0.89496163
#> sample_2   0.3861270  0.02083123  0.007586777  0.93084403 -0.68474994
#> sample_3   0.8769131 -1.14890394 -1.137612756 -1.43724674 -0.49414348
#> sample_4  -1.2893009  0.37511557 -0.556918391  0.30242663  0.22226647
#>              gene_115    gene_116    gene_117     gene_118    gene_119
#> sample_1   0.06730444 -0.16267634 -0.82731017  1.876505621  0.76644020
#> sample_2   0.33740151 -0.41213770  0.93426113  1.840316741 -0.70481966
#> sample_3   0.84080181  0.79153412 -0.16884895  0.612722104 -0.77115892
#> sample_4  -0.96216696  0.05203237  0.90004038 -0.394900880  0.95414600
#>               gene_120     gene_121    gene_122    gene_123    gene_124
#> sample_1   0.979956696  1.321780992 -1.11971083  0.51459982 -1.50909984
#> sample_2   0.008510312  2.034189886 -1.34168607  1.15897918 -0.20320896
#> sample_3   0.888628993  0.013214477  0.22533951 -0.72991521 -1.22248707
#> sample_4  -0.588214348 -1.476584527 -1.37777577 -1.34567231 -0.73663796
#>              gene_125   gene_126   gene_127   gene_128    gene_129    gene_130
#> sample_1   1.53274148  0.4291474  0.1221034 -1.1380124 -0.55801513  1.05253854
#> sample_2  -0.37802855  1.7361110 -0.8452478 -0.9615715  1.01749105 -1.49605374
#> sample_3   0.40680517 -0.7510122 -0.1621165  0.3520101 -0.28905830  0.10466223
#> sample_4  -0.47011150  1.3806093  1.6750109  1.1769066 -0.14889834 -0.17782336
#>               gene_131    gene_132    gene_133    gene_134    gene_135
#> sample_1   0.677683644  0.03849955 -0.35638119  0.78284410  0.80441162
#> sample_2  -1.184818730  0.63023437  2.10125251 -0.61373681 -1.63463827
#> sample_3   0.720186531 -0.61104608 -1.10691407  0.53480333  0.73606797
#> sample_4   0.803121858 -0.41579535  1.21248959  1.24036003  0.68567588
#>               gene_136   gene_137    gene_138    gene_139   gene_140
#> sample_1  -1.900060823  0.9357843 -0.30905150  0.26306668 -1.7905919
#> sample_2  -0.010441117 -0.6565061 -0.66953344 -0.47858903  1.3194563
#> sample_3  -1.222501574  1.0214153  0.46516516  0.79047270 -0.1302648
#> sample_4  -0.026798678  0.3095805  0.24986433 -1.35646087  0.5993798
#>              gene_141    gene_142    gene_143    gene_144     gene_145
#> sample_1  -0.78825884 -1.13302167  0.36365257 -0.28588791  0.517669134
#> sample_2   0.63656276  0.51432778 -1.75137511  0.89359752  0.223038372
#> sample_3  -0.93028533 -0.36485100  0.15387249  0.41315482  2.480823360
#> sample_4   0.00864779  0.09071661 -0.65705741 -0.48178225  0.017757879
#>              gene_146    gene_147    gene_148     gene_149    gene_150
#> sample_1  -0.10290867 -0.97406959  1.27067230  0.960864787  0.76872137
#> sample_2   0.58081659 -0.17782142  0.74096671 -0.997443079 -2.93897756
#> sample_3  -2.17995674  0.42087458 -0.35752832 -0.646861514 -0.05014180
#> sample_4  -0.85953576  1.35770498  1.21506118  1.453912917 -0.08591198
#>             gene_151    gene_152   gene_153    gene_154    gene_155    gene_156
#> sample_1   1.0359308 -0.47388707 -1.2753349 -0.30562067  2.21176949 -1.04166838
#> sample_2   0.7190157 -0.69800504 -1.8941258  0.07629925  0.87530850  0.45382739
#> sample_3   0.4169428 -0.63258754  1.1501467 -0.23547591 -1.64310739 -1.50338215
#> sample_4  -0.6175688 -0.21826025 -1.3260052 -2.36220890 -1.40996955  0.25439714
#>             gene_157    gene_158    gene_159   gene_160    gene_161   gene_162
#> sample_1  -1.1465239 -1.67532730  1.52593866  0.5541855  1.99311026 -0.1541207
#> sample_2  -0.8507169  0.56620161  1.15221195 -0.7561974 -0.48925833 -1.1660523
#> sample_3  -2.0505848 -0.75319823 -0.13414196  1.0057828  2.16718680  2.3225565
#> sample_4   0.2958703  0.05678979 -0.10265486  1.9519230  0.87856596 -1.3066169
#>               gene_163    gene_164    gene_165    gene_166    gene_167
#> sample_1   2.564408338  1.06199914  1.14269488  1.12383884 -0.39700149
#> sample_2  -0.479668950  0.11534822 -1.76804841 -1.40763892  0.70917846
#> sample_3  -1.020423391  0.04881444 -0.77188863 -0.78523507 -0.72660303
#> sample_4   0.090344272  0.90235804  0.54942132 -0.86983920 -0.03921465
#>             gene_168   gene_169    gene_170      gene_171   gene_172   gene_173
#> sample_1  -0.8232612 -0.5788846  1.76378938  0.1329921464  0.3764993  1.1387077
#> sample_2  -1.2408429 -0.3683273  0.46208009 -0.3228331007 -1.2872148 -1.0300402
#> sample_3   0.6818780 -0.2298433 -1.51060172 -0.5837276871 -2.0229185  0.4035047
#> sample_4  -0.5184517 -0.9118496  0.15031364  0.4403554485  1.3197821  0.2465611
#>              gene_174    gene_175   gene_176    gene_177    gene_178
#> sample_1   1.24126308  0.61209094 -0.4293801  1.36046133 -0.07085743
#> sample_2   1.51408932  0.34690359  1.7794415  0.38663092 -0.91869524
#> sample_3   0.55001555  0.02835712  0.8931650 -0.37655550  0.60588481
#> sample_4   0.94182660 -0.34215133 -0.2761468  0.38365678  1.44691022
#>               gene_179    gene_180    gene_181     gene_182    gene_183
#> sample_1  -0.272153684 -2.44668003  0.06548664 -1.098508902 -0.63317818
#> sample_2  -1.584336488 -0.08405889 -2.08507089  0.003567992 -0.35577082
#> sample_3  -0.004874726 -0.52079637 -0.63901860 -0.635894137  0.10658698
#> sample_4   1.733900291  0.45642938  0.70750349  2.063789224  0.02391059
#>              gene_184     gene_185    gene_186    gene_187    gene_188
#> sample_1  -2.06365445  2.648932029 -1.15339839 -0.34063788  0.78636258
#> sample_2   1.14635975 -0.221188446  1.01817902 -0.26371930  1.65854230
#> sample_3   1.17691425  0.447391153  2.27295477  0.13605821 -1.99903913
#> sample_4   0.24594946  0.272059145 -0.99245858 -0.02757795  2.22284516
#>             gene_189    gene_190   gene_191    gene_192    gene_193    gene_194
#> sample_1  -1.2705131  0.54214155  0.0751059  0.55851442  0.41540640 -1.45229977
#> sample_2  -0.7740868 -0.92393788 -0.2755334 -0.59339969 -0.12228589  1.17978425
#> sample_3  -0.4205009 -0.37840740  1.2207748 -1.54103029 -0.31031012 -0.02010818
#> sample_4   0.1555039 -0.86911632 -1.1744894 -1.75967731  0.05836159  1.16451986
#>              gene_195    gene_196    gene_197     gene_198   gene_199
#> sample_1   0.94120612 -0.33893587 -0.07557425  0.040204392  0.1243011
#> sample_2   0.64103737 -0.62958851 -0.80773497 -0.860489929 -2.1692387
#> sample_3  -2.39020034  0.88986536 -1.48281332  0.445750348  1.3697759
#> sample_4   0.33762789 -1.05451872  0.66076911 -0.823409004  0.4370337
#>              gene_200    gene_201    gene_202   gene_203   gene_204    gene_205
#> sample_1  -0.99843255  1.23339006  0.34042449 -0.4727025  0.7087531 -1.52895871
#> sample_2  -1.37583652 -0.49313247 -0.58165203 -0.1672293  0.4859931 -1.33339580
#> sample_3  -0.02011003 -0.10921759  0.26466174  0.3038483 -0.1833885  0.55964967
#> sample_4   0.37238811 -1.64674913 -1.92355227  0.3808330  1.3757053  0.82584873
#>              gene_206   gene_207    gene_208    gene_209    gene_210   gene_211
#> sample_1   0.23742535 -1.3128142  0.74702859 -1.56251843  0.07105336 -0.6395348
#> sample_2  -0.26196562  0.6523863  0.74885497  0.89656029  1.48930042 -0.6594035
#> sample_3  -0.18655384 -0.8122754 -1.64058167  0.50792248  1.75433696  0.5924002
#> sample_4  -0.41611500  0.9821496  0.24218073 -0.93792685  1.50585850 -1.0658512
#>              gene_212   gene_213    gene_214   gene_215    gene_216    gene_217
#> sample_1  -0.84519574  0.6752447  1.15337579 -1.6865047 -0.90281495  1.31763370
#> sample_2   0.53728318  0.7468031  1.89631708 -2.0600707  0.06454387 -0.26514740
#> sample_3   1.01671329  0.1216206 -1.07806726 -1.1435657 -0.52964368 -0.68127316
#> sample_4   0.16227053  0.2604704 -1.48820199  1.4151839  0.56250751  0.64205574
#>              gene_218     gene_219    gene_220    gene_221    gene_222
#> sample_1   1.10018974  1.203767839 -1.43127078  1.38291086  0.00312594
#> sample_2  -0.44734453 -1.410700927 -0.50641888 -0.26976184 -1.08515492
#> sample_3  -0.20244756  1.684495721 -1.03377324 -0.15597667 -0.04640064
#> sample_4  -0.89053264 -0.116537517 -0.94197475  1.11582792  0.47013513
#>              gene_223    gene_224    gene_225    gene_226   gene_227
#> sample_1  -0.07788682  0.44142823  0.12892290 -0.83021426 -0.5035929
#> sample_2   0.36215913 -0.33567214  1.36380450 -0.71152414  0.6621788
#> sample_3  -0.95362873  0.41626080  0.11402961  0.06391875 -0.9193322
#> sample_4   0.86061271 -0.07039665 -0.61318021 -0.33671496 -0.2158402
#>              gene_228    gene_229    gene_230    gene_231    gene_232
#> sample_1  -1.19364118 -0.75172332  1.45584140 -0.82860353  0.28977446
#> sample_2   0.29113022  0.19795800 -1.20356611 -0.03981704  0.68698246
#> sample_3   0.90133529 -0.79772830  0.66822120  0.15521430  0.12868809
#> sample_4   0.62113229 -1.28402652 -1.30009244 -0.37676946  0.10374865
#>             gene_233    gene_234     gene_235    gene_236    gene_237
#> sample_1  -0.4800535 -0.60482935  1.460110180  0.14967935 -1.43332110
#> sample_2   0.7052670  0.99144168  1.144248971 -1.23891024  2.65489833
#> sample_3  -1.5330655  0.20236067 -0.717538653  0.36169476  1.39900429
#> sample_4  -0.7035623  1.49741394 -0.302826878 -1.37683133  0.88317017
#>              gene_238    gene_239    gene_240   gene_241    gene_242
#> sample_1  -0.01030332 -0.21223603 -0.90634018 -2.1021525  1.89336046
#> sample_2  -0.15691719 -0.42349012 -0.19838706 -0.8948024  0.90426912
#> sample_3   0.37269896 -1.56564429 -0.05169454  0.5140821  0.54989952
#> sample_4   0.67142028  1.23002238  1.64546788 -0.1829260 -1.37925781
#>              gene_243    gene_244    gene_245    gene_246    gene_247
#> sample_1  -0.96812584 -0.10260304  0.23995957  0.06089889 -2.17757603
#> sample_2   0.07964921 -1.25882722  1.02568511 -0.73077860 -0.19014551
#> sample_3   0.86781691  0.68436008 -0.16267998 -1.78436472 -1.03714557
#> sample_4   1.25060060 -1.87623677 -0.36878757 -2.12929731  0.33764603
#>              gene_248   gene_249     gene_250
#> sample_1  -0.11786014  0.1122948  0.007886198
#> sample_2   0.52886469  0.5502105  0.549684337
#> sample_3   0.83014772  0.6073469 -0.122186361
#> sample_4  -1.57457485 -0.1963194  0.915060108
#>  [ reached getOption("max.print") -- omitted 16 rows ]