Check validity of new Omics*-class objects
Source: R/createClass_validOmics.R

These functions check the validity of new objects created in the OmicsSurv, OmicsReg, and OmicsCateg classes.

valid_OmicsSurv(object)

valid_OmicsReg(object)

valid_OmicsCateg(object)

Arguments

object
An object potentially of class OmicsSurv, OmicsReg, or OmicsCateg.

Value

TRUE if the object is a valid object, else an error message with the rule broken.

Details

We have currently written checks to make sure the dimensions of the mass spectrometry or bio-assay data frame and response matrices or vectors match. Other checks should be added in response to user feedback during or after beta testing. FIX THIS.

OmicsSurv

Valid OmicsSurv objects will have two response vectors: a vector of the most recently recorded follow-up times and a logical vector if that time marks a death or event (TRUE: observed event; FALSE: right-censored observation).

OmicsReg and OmicsCateg

Valid OmicsReg and OmicsCateg objects with have one response vector of continuous (numeric) or categorial (factor) observations, respectively.