Calculate the \(p\)-values from an optimal mixture of Weibull Extreme Value distributions for supervised PCA
Source: R/superPC_optimWeibull_pValues.R

Calculate the \(p\)-values of test statistics from a mixture of two Weibull Extreme Value distributions.

weibullMix_pValues(tScore_vec, pathwaySize_vec, optimParams_vec)

Arguments

tScore_vec
A vector of the maximum absolute \(t\)-scores for each pathway (returned by the pathway_tScores function) when under the alternative model.

pathwaySize_vec
A vector of the number of genes in each pathway.

optimParams_vec
The NAMED vector of optimal Weibull Extreme Value mixture distribution parameters returned by the weibullMix_optimParams function.

Value

A named vector of the estimated raw \(p\)-values for each gene pathway.

Details

The likelihood function is equation (4) in Chen et al (2008): a mixture of two Gumbel Extreme Value probability density functions, with mixing proportion \(p\). Within the code of this function, the values mu1, mu2 and s1, s2 are placeholders for the mean and precision, respectively.

See https://doi.org/10.1093/bioinformatics/btn458 for more information.

See also

Examples

  # DO NOT CALL THIS FUNCTION DIRECTLY.
  # Use superPCA_pVals() instead.