Wikipathways Homosapiens Gene Symbols

A pathwayCollection object containing the homosapiens pathways list from Wikipathways (https://www.wikipathways.org/).

wikipwsHS_Symbol_pathwayCollection

Format

A pathwayCollection list of three elements:

pathways : A named list of 457 character vectors. Each vector contains the Gene Symbols of the individual genes within that pathway as a vector of character strings. The names are the shorthand pathway names.

TERMS : A character vector of length 457 containing the shorthand names of the gene pathways.

description : A character vector of length 457 containing the full names of the gene pathways.

Source

Dr. Alexander Pico, Wikipathways

Details

This pathwayCollection was sent to us from Dr. Alexander Pico at the Gladstone Institute (https://gladstone.org/our-science/people/alexander-pico).

This pathway collection was translated from EntrezIDs to HGNC Symbols with the script convert_EntrezID_to_HGNC_Ensembl.R in scripts.