Write a pathwayCollection Object to a .gmt File

Write a pathwayCollection object as a pathways list file in Gene Matrix Transposed (.gmt) format.

write_gmt(pathwayCollection, file)

Arguments

pathwayCollection
A pathwayCollection list of pathways. This list contains the following two or three elements:

pathways : A named list of character vectors. Each vector contains the names of the individual genes within that pathway as a vector of character strings. Genes can be represented by HGNC gene symbols, Entrez IDs, Ensembl IDs, GO terms, etc.

TERMS : A character vector the same length as the pathways list with the proper names of the pathways.

description : An optional character vector the same length as the pathways list with a note on that pathway (such as a url to the description of the pathway). If this element of the pathwayCollection is NULL, then the file will be written with "" (the empty character string) as its second field in each line.

file
Either a character string naming a file or a connection open for writing. File names should end in .gmt for clarity

Value

NULL. Output written to the file path specified.

Details

See the Broad Institute's "Data Formats" page for a description of the Gene Matrix Transposed file format: https://software.broadinstitute.org/cancer/software/gsea/wiki/index.php/Data_formats#GMT:_Gene_Matrix_Transposed_file_format_.28.2A.gmt.29

See also

Examples

# NOT RUN {
  # Toy pathway set
  toy_pathwayCollection <- list(
    pathways = list(
      c("C1orf27", "NR5A1", "BLOC1S4", "C4orf50"),
      c("TARS2", "DUSP5", "GPR88"),
      c("TRX-CAT3-1", "LINC01333", "LINC01499", "LINC01046", "LINC01149")
    ),
    TERMS = c("C-or-f_paths", "randomPath2", "randomLINCs"),
    description = c("these are", "totally made up", "pathways")
  )
  class(toy_pathwayCollection) <- c("pathwayCollection", "list")
  print(toy_pathwayCollection)

  write_gmt(toy_pathwayCollection, file = "example_pathway.gmt")
# }