An S4 class for categorical responses within an OmicsPathway object

This creates the OmicsCateg class which extends the OmicsPathway master class.

Slots

assayData_df

An \(N \times p\) data frame with named columns.

pathwayCollection

A list of known gene pathways with three or four elements:

pathways : A named list of character vectors. Each vector contains the names of the individual genes within that pathway as a vector of character strings. The names contained in these vectors must have non-empty overlap with the column names of the assayData_df data frame. The names of the pathways (the list elements themselves) should be the a shorthand representation of the full pathway name.

TERMS : A character vector the same length as the pathways list with the proper names of the pathways.

description : An optional character vector the same length as the pathways list with additional information about the pathways.

setsize : A named integer vector the same length as the pathways list with the number of genes in each pathway. This list item is calculated during the creation step of a CreateOmics function call.

response

A factor vector of length \(N\): the dependent variable of a generalized linear regression exercise. Currently, we support binary factors only. We expect to extend support to n-ary responses in the next package version.

See also