Subset Pathway-Specific Data

Given an Omics object and the name of a pathway, return the -omes in the assay and the response as a (tibble) data frame.

SubsetPathwayData(object, pathName, ...)

# S4 method for OmicsPathway
SubsetPathwayData(object, pathName, ...)

Arguments

object
An object of class OmicsPathway, or an object extending this class.

pathName
The name of a pathway contained in the pathway collection in the object.

...
Dots for additional internal arguments (currently unused).

Value

A data frame of the columns of the assay in the Omics object which are listed in the specified pathway, with a leading column for sample IDs. If the Omics object has response information, these are also included as the first column(s) of the data frame, after the sample IDs. If you have the suggested tidyverse package suite loaded, then this data frame will print as a tibble. Otherwise, it will print as a data frame.

Details

This function subsets the assay by the matching gene symbols or IDs in the specified pathway.

Examples

  data("colonSurv_df")
  data("colon_pathwayCollection")

  colon_Omics <- CreateOmics(
    assayData_df = colonSurv_df[, -(2:3)],
    pathwayCollection_ls = colon_pathwayCollection,
    response = colonSurv_df[, 1:3],
    respType = "survival"
  )

#> 
#>   ======  Creating object of class OmicsSurv  =======

#> The input pathway database included 676 unique features.

#> The input assay dataset included 656 features.

#> Only pathways with at least 3 or more features included in the assay dataset are
#>   tested (specified by minPathSize parameter). There are 15 pathways which meet
#>   this criterion.

#> Because pathwayPCA is a self-contained test (PMID: 17303618), only features in
#>   both assay data and pathway database are considered for analysis. There are 615
#>   such features shared by the input assay and pathway database.

  SubsetPathwayData(
    colon_Omics,
    "KEGG_RETINOL_METABOLISM"
  )

#> # A tibble: 250 x 48
#>    sampleID EventTime EventObs   RPE65  CYP3A5 UGT2B28 CYP4A11  CYP3A4    RDH8
#>    <chr>        <dbl> <lgl>      <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 subj1       64.9   FALSE    -0.721   1.05     0.598   0.844  1.03   -1.42  
#>  2 subj2       59.8   FALSE     2.25    1.74     2.23    0.724  2.73    1.05  
#>  3 subj3       62.4   FALSE     0.236  -0.651   -1.01   -1.88  -0.335  -1.13  
#>  4 subj4       54.5   FALSE     0.221   0.541    1.40    3.00   1.54   -0.433 
#>  5 subj5       46.3   TRUE      0.157   0.0710  -1.40    1.90   0.299  -1.43  
#>  6 subj6       55.9   FALSE    -0.744  -0.912   -0.803   0.645 -0.894   1.91  
#>  7 subj7       58.0   FALSE    -0.480  -0.543   -0.272   1.64  -0.0453  1.64  
#>  8 subj8       54.0   FALSE    -0.0133 -1.08    -0.746   0.268 -0.636   0.339 
#>  9 subj9        0.427 TRUE     -0.307  -0.149    0.825  -0.224 -0.453  -0.0897
#> 10 subj10      41.4   FALSE    -0.300   1.81     0.430   0.356  0.774   0.955 
#> # ... with 240 more rows, and 39 more variables: DHRS3 <dbl>, CYP2C18 <dbl>,
#> #   ADH1B <dbl>, ADH1C <dbl>, ADH4 <dbl>, ADH5 <dbl>, DGAT2 <dbl>, ADH1A <dbl>,
#> #   RDH10 <dbl>, CYP26A1 <dbl>, CYP2C9 <dbl>, CYP2C19 <dbl>, CYP2C8 <dbl>,
#> #   CYP2B6 <dbl>, UGT2A3 <dbl>, CYP2A13 <dbl>, UGT1A1 <dbl>, CYP26B1 <dbl>,
#> #   DHRS9 <dbl>, DGAT1 <dbl>, ALDH1A1 <dbl>, RETSAT <dbl>, RDH12 <dbl>,
#> #   LRAT <dbl>, RDH11 <dbl>, CYP2A6 <dbl>, CYP2A7 <dbl>, CYP1A1 <dbl>,
#> #   CYP1A2 <dbl>, ADH7 <dbl>, ADH6 <dbl>, ALDH1A2 <dbl>, CYP3A43 <dbl>,
#> #   RDH16 <dbl>, RDH5 <dbl>, PNPLA4 <dbl>, UGT2B4 <dbl>, UGT2B17 <dbl>,
#> #   UGT2B15 <dbl>