Adaptive, elastic-net, sparse principal component analysis

A function to perform adaptive, elastic-net, sparse principal component analysis (AES-PCA).

aespca(X, n, d = 1, lambda = 1e-04, type = c("predictor", "Gram"),
  corr = FALSE, max.iter = 10, adaptive = TRUE, para = NULL)

Arguments

X
A pathway design matrix: either the data matrix itself (when type = "predictor") or the Gram matrix (when type = "Gram"). The data matrix should be \(n x p\), where \(n\) is the sample size and \(p\) is the number of variables included in the pathway.

n
The sample size. Needed when X is the Gram matrix (for computing BIC). Due to a design error, it is a required argument even when X is not a Grammian. FIX THIS.

d
The number of PCs to extract from the pathway. Defaults to 1.

lambda
The ridge regression penalty. Defaults to \(10^{-4}\).

type
Is X a pathway design matrix or a Grammian? Defaults to both: c("predictor", "Gram"). FIX THIS TOO.

corr
If type = "Gram", is the matrix X actually a correlation matrix? Defaults to FALSE.

max.iter
The maximum number of times an internal while() loop can make calls to the lars.lsa() function. Defaults to 10.

adaptive
Internal argument of the lars.lsa() function. Defaults to TRUE.

para
Internal argument of the lars.lsa() function. Defaults to NULL.

Value

What does the function return? DOCUMENT THIS.

Details

A thorough explination of how the function works. DOCUMENT THIS.

Examples

  # DO NOT CALL THIS FUNCTION DIRECTLY.
  # Call this function through AESPCA_pVals() instead.