Access and Edit Assay or pathwaySet Values in Omics* Objects
Source: R/accessClass_OmicsPath.R

"Get" or "Set" the values of the assayData_df or pathwaySet slots of an object of class OmicsPathway or a class that extends this class (OmicsSurv, OmicsReg, or OmicsCateg).

getAssay(object, ...)

getAssay(object) <- value

getPathwaySet(object, ...)

getPathwaySet(object) <- value

# S4 method for OmicsPathway
getAssay(object, ...)

# S4 method for OmicsPathway
getAssay(object) <- value

# S4 method for OmicsPathway
getPathwaySet(object, ...)

# S4 method for OmicsPathway
getPathwaySet(object) <- value

Arguments

object
An object of or extending OmicsPathway-class: that class, OmicsSurv-class, OmicsReg-class, or OmicsCateg-class.

...
Dots for additional internal arguments (currently unused).

value
The replacement object to be assigned to the specified slot.

Value

The "get" functions return the objects in the slots specified: getAssay returns the assayData_df data frame object and getPathwaySet returns the pathwaySet list object. These functions can extract these values from any valid OmicsPathway, OmicsSurv, OmicsReg, or OmicsCateg object.

The "set" functions enable the user to edit or replace objects in the assayData_df or pathwaySet slots for any OmicsPathway, OmicsSurv, OmicsReg, or OmicsCateg objects, provided that the new values do not violate the validity checks of their respective objects. See "Details" for more information.

Details

These functions can be useful to set or extract the assay data or gene set list from an Omics*-class object. However, we recommend that users simply create a new, valid Omics* object instead of modifying an existing one. The validity of edited objects is checked with the valid_OmicsSurv, valid_OmicsCateg, or valid_OmicsReg functions.

See also

Examples

# NOT RUN {
  getAssay(colon_OmicsSurv)
  getPathwaySet(colon_OmicsSurv)

  getAssay(colon_OmicsSurv) <- newAssay_df
  getPathwaySet(colon_OmicsSurv) <- new_pathwaySet
# }