Calculate the \(p\)-values from a mixture of Weibull Extreme Value Distributions for supervised PCA
Source: R/superPC_pathway_pValues.R

Calculate pathway-specific \(p\)-values for supervised PCA and their associated False Discovery Rates (FDR).

pathway_pValues(optimParams_vec, max_tScores_vec, genelist_ls,
  FDRadjust = TRUE, multTestProc = "BH")

Arguments

optimParams_vec
A named vector of the estimated values for the parameters which minimize the likelihood as returned by the function weibullMix_optimParams.

max_tScores_vec
A vector of the maximum absolute \(t\)-scores for each pathway when under the alternative model (the response vector as is).

genelist_ls
A list of three elements:

pathways : A list of character vectors such that each vector contains the ID numbers (as a character) of the individual genes within that pathway as a vector of character strings.

TERMS : A character vector containing the names of the gene pathways.

setsize : A named integer vector containing the number of genes in each gene pathway.

FDRadjust
Should the \(p\)-values be adjusted for multiple comparisons? Defaults to TRUE.

multTestProc
If the \(p\)-values should be adjusted, which procedure should be used? Options are passed to the adjustRaw_pVals function. Specify multiple procedures via c(...). Defaults to "BH".

Value

A data frame with columns for the pathway names, pathway set sizes, raw pathway \(p\)-values, and a column of FDR-adjusted \(p\)-values for each adjustment method specified.

Details

This function takes in the optimal parameters returned by the weibullMix_optimParams function, the maximum \(t\)-scores for each gene pathway, and the list of gene pathway information. This function will calculate the \(p\)-value for each \(t\)-score given the Gumbel Extreme Value mixture distribution parametrized by the values returned by the weibullMix_optimParams function. If requested, this function will also calculate the FDR associated with all pathway \(p\)-values via requested FDR-adjustment procedure. The default procedure is the Benjamini & Hochberg (1995) step-up FDR-controlling procedure, but any procedure implemented in the adjustRaw_pVals function is available.

Examples

  # DO NOT CALL THIS FUNCTION DIRECTLY.
  # Use superPCA_pVals() instead