Parametric bootstrap and non-parametric permutations of a response vector or matrix

Create a random parametric bootstrap sample or a permutation of the input response vector or matrix (for survival outcomes).

SampleResponses(response_vec, event_vec = NULL,
  respType = c("survival", "regression", "categorical"),
  parametric = TRUE)

SampleSurv(response_vec, event_vec, parametric = TRUE)

SampleReg(response_vec, parametric = TRUE)

SampleCateg(response_vec, parametric = TRUE)

Arguments

response_vec
The dependent vector to sample from. For survival response, this is the vector of event times. For regression or n-ary classification, this is the vector of responses.

event_vec
The death / event observation indicator vector for survival response. This is coded as 0 for a right-censoring occurence and 1 for a recorded event.

respType
What type of response has been supplied. Options are "none", "survival", "regression", and "categorical". Defaults to "none" to match the default response = NULL value.

parametric
Should the random sample be taken using a parametric bootstrap sample? Defaults to TRUE.

Value

If parametric = FALSE, a permutation of the supplied response is returned (for AES-PCA). If parametric = TRUE, we return a parametric bootstrap sample of the response.

Details

The distributions (for parametric = TRUE) are Weibull for survival times, Normal for regression response, and n-ary Multinomial for categorical response. Distributional parameters are estimated with their maximum likelihood estimates. When parametric = FALSE, the response vector or survival matrix is randomly ordered by row. This option should only be used when called from the AES-PCA method.

Examples

  # DO NOT CALL THESE FUNCTIONS DIRECTLY.
  # Use AESPCA_pVals() or SuperPCA_pVals() instead

  data("colon_pathwayCollection")
  data("colonSurv_df")

  SampleResponses(
    response_vec = colonSurv_df$OS_time,
    event_vec = colonSurv_df$OS_event,
    respType = "survival"
  )

#> $response_vec
#>   [1]  34.0749565   3.6784209 199.8912139 158.6626779   6.8984210 376.8397640
#>   [7]  13.8435509 186.1716092  17.0208140   0.8535502 191.7464613  13.9210846
#>  [13] 304.9216947   1.6010998  27.6854503  48.1316873 195.8836185   1.3171466
#>  [19]  28.0757536 159.8194795  50.9875412  15.4199048 108.3042141  12.6268193
#>  [25]  72.2352444  87.1178012  27.3710683   1.0561437  73.0434995  31.1264310
#>  [31] 231.7256753 416.4933773  12.1043445 260.8695004  27.7661878 205.4192093
#>  [37] 304.9255630   5.6241431  16.6454315  19.5388443  18.9594597 338.7218774
#>  [43]  25.2911933 195.4675158  97.8495616 110.0705093 233.5342235  92.6218503
#>  [49] 129.3807719 197.3627670  67.5645116  46.7738726  46.5496049 749.6508860
#>  [55]   0.4356056  53.0298018 635.9921923 210.5163166  46.0107539  93.1713553
#>  [61]  46.9874536  41.1340108 176.0432251  16.0191265  29.9738007  70.5325831
#>  [67]  46.2846019 136.7698406  74.8972328 152.1355897  11.5768951  36.0240396
#>  [73] 111.2171461 112.3344322 431.1419388  29.2627862  72.0756422 484.8356017
#>  [79] 284.7395393  16.3405419 122.0220301  85.1523040  98.3205710  40.9825531
#>  [85]   5.9001042  73.2486337   1.1221074   4.1641478   0.7107276  99.4079845
#>  [91] 254.1991919   2.9464282   6.8072680  71.2021687 200.0374111 195.5287478
#>  [97] 162.7468954 111.6430209  96.2105065 144.3549661  16.3067341  33.7242443
#> [103]  23.5785130  52.3437194  99.2293865  89.0068283 164.7180779  16.0913403
#> [109] 570.7526180 179.5076271  47.2298889   7.7882545  34.0401819 390.3674926
#> [115]  40.5365149 189.4703965   4.1135735  47.2669211 131.1211889  54.3502816
#> [121]  34.4544429   0.2799212  63.3662271 171.6177554  42.0212809  33.0739473
#> [127] 174.6078773 309.3083045 174.0383265 190.4587129  72.0482221 180.7570960
#> [133]   8.5339844  44.2837204   8.4189726 166.2233025   0.8243824  38.9188156
#> [139]   6.3980586   3.0288678   5.0977145  27.2589195 137.8031745  21.1886625
#> [145]  46.0440709 113.1071766   6.7557898 144.0001851 397.8514876  26.4638397
#> [151]  63.4653143  32.0663483   0.7460259  74.1892289 730.4837673  70.0278374
#> [157]  21.2185320  49.7952841 180.7801298  34.2847443  44.3064273  20.2901593
#> [163]   8.2328834  39.9797813  56.5448352  49.1338978  95.3209868 231.3469872
#> [169]  57.8689915  17.8021598   5.2333377  17.3697800  97.6948892  27.7458273
#> [175]  19.8888704 326.4084254  85.6159888 104.0250338 147.5702420  34.2122119
#> [181] 430.2970834  12.9549958   7.4735955 194.8242255   3.3791126  41.1220608
#> [187]  26.7136949  85.3050226   5.0729877 209.2132540  12.7731723   2.4596495
#> [193] 150.0337995  31.8166393 190.5759853  11.0733204 140.9020101 120.3122461
#> [199]  56.9314154  80.3412408  74.6069506  46.7686974   3.3155918  88.3705217
#> [205] 257.3610087   9.6298745 101.0036141  22.1454387  88.0545877 176.5105188
#> [211]  87.4190134 247.5434673  39.1632915  35.8993548  49.5823182  22.6589552
#> [217]   1.7975882  54.7929943 214.0107979  27.5703863  19.9221883 218.4752494
#> [223]  54.6671200 122.9268792  25.7106076 111.1061815  67.4403586  23.3501975
#> [229]  23.3988386  88.6107141 395.3270738 313.3483525  63.1840243  60.4919750
#> [235]  73.2274645  77.8000796  50.2642444  18.8867090   4.6144069   6.5865475
#> [241]  49.4517867  25.6309002 405.7193431 129.0836635  12.0300267  21.3933991
#> [247]  24.6817219   2.1672012  37.6852923  29.6772354
#> 
#> $event_vec
#>   [1]  TRUE FALSE  TRUE  TRUE FALSE FALSE FALSE  TRUE FALSE FALSE  TRUE  TRUE
#>  [13]  TRUE FALSE FALSE  TRUE  TRUE FALSE FALSE  TRUE  TRUE FALSE FALSE FALSE
#>  [25]  TRUE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE FALSE
#>  [37]  TRUE  TRUE  TRUE FALSE FALSE  TRUE  TRUE  TRUE  TRUE FALSE  TRUE FALSE
#>  [49]  TRUE FALSE  TRUE  TRUE  TRUE FALSE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [61] FALSE  TRUE  TRUE  TRUE FALSE FALSE  TRUE FALSE FALSE  TRUE FALSE  TRUE
#>  [73]  TRUE FALSE  TRUE FALSE FALSE  TRUE FALSE FALSE FALSE  TRUE  TRUE  TRUE
#>  [85] FALSE FALSE  TRUE  TRUE  TRUE  TRUE  TRUE FALSE FALSE FALSE  TRUE FALSE
#>  [97] FALSE  TRUE FALSE FALSE FALSE  TRUE FALSE  TRUE FALSE FALSE  TRUE  TRUE
#> [109]  TRUE  TRUE  TRUE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE  TRUE FALSE
#> [121]  TRUE  TRUE FALSE FALSE FALSE FALSE  TRUE  TRUE  TRUE FALSE  TRUE FALSE
#> [133] FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE
#> [145] FALSE  TRUE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE
#> [157]  TRUE  TRUE FALSE  TRUE  TRUE  TRUE  TRUE FALSE FALSE  TRUE  TRUE FALSE
#> [169]  TRUE FALSE FALSE  TRUE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE
#> [181]  TRUE FALSE FALSE  TRUE  TRUE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE
#> [193] FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE  TRUE FALSE FALSE
#> [205]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE FALSE  TRUE FALSE FALSE FALSE FALSE
#> [217] FALSE  TRUE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE  TRUE  TRUE FALSE
#> [229] FALSE FALSE  TRUE FALSE  TRUE FALSE FALSE  TRUE  TRUE  TRUE FALSE  TRUE
#> [241]  TRUE  TRUE  TRUE  TRUE FALSE FALSE FALSE FALSE  TRUE FALSE
#>