Read a .gmt file in as a pathwaySet object
Source: R/read_gmt.R

Read a pathways list file in Gene Matrix Transposed (.gmt) format, with special performance consideration for large files. Present this object as a pathwaySet object.

read_gmt(file, description = FALSE, delim = "\t")

Arguments

file
A path to a file or a connection. This file must be a .gmt file, otherwise input will likely be nonsense. See the "Details" section for more information.

description
Should the "description" field (the second field in the .gmt file on each line) be included in the output? Defaults to FALSE.

delim
The .gmt delimiter. As proper .gmt files are tab delimited, this defaults to "\t".

Value

A pathwaySet list of pathways. This list has three elements:

pathways : A named list of character vectors. Each vector contains the names of the individual genes within that pathway as a vector of character strings.

TERMS : A character vector the same length as the pathways list with the proper names of the pathways.

description : (OPTIONAL) A character vector the same length as the pathways list with a note on that pathway (for the .gmt file included with this package, this field contains hyperlinks to the MSigDB description card for that pathway). This field is included when description = TRUE.

Details

This function uses R's readChar function to improve character input performance over readLines (and far improve input performance over scan).

See the Broad Institute's "Data Formats" page for a description of the Gene Matrix Transposed file format: https://software.broadinstitute.org/cancer/software/gsea/wiki/index.php/Data_formats#GMT:_Gene_Matrix_Transposed_file_format_.28.2A.gmt.29

See also

Examples

# NOT RUN {
  # If you have installed the package:
  data_path <- system.file("extdata", "c2.cp.v6.0.symbols.gmt",
                            package = "pathwayPCA", mustWork = TRUE)
  geneset_ls <- read_gmt(data_path)

  # If you are using the development version from GitHub:
  geneset_ls <- read_gmt("inst/extdata/c2.cp.v6.0.symbols.gmt")
# }