Train a supervised PCA model

Computes feature scores for \(p_{path}\) features of a pathway via a linear model fit.

superpc.train(data, type = c("survival", "regression", "categorical"),
  s0.perc = NULL)

Arguments

data
A list of test data:

x : A "tall" pathway data frame (\(p_{path} \times N\)).

y : A response vector corresponding to type.

censoring.status : If type = "survival", the censoring indicator (\(1 - \) the observed event indicator. Otherwise, NULL.

featurenames : A character vector of the measured -Omes in x.

type
What model relates y and x? Options are "survival", "regression", or "categorical".

s0.perc
A stabilization parameter on the interval \([0,1]\). This is an internal argument to each of the called functions. The default value is NULL to ensure an appropriate value is determined internally.

Value

A list containing:

feature.scores : The scaled \(p\)-dimensional score vector: each value has been divided by its respective standard deviation plus epsilon (governed by s0.perc). NA values returned by the logistic model are replaced with 0.

type : The argument for type.

s0.perc : The user-supplied value of s0.perc, or the internally-calculated default value from the chosen model.

call : The output of match.call for the user- supplied function arguments.

Details

This function is a switch call to coxTrain_fun (for type = "survival"), olsTrain_fun (for type = "regression"), or glmTrain_fun (for type = "categorical").

See also

Examples

  # DO NOT CALL THIS FUNCTION DIRECTLY.
  # Use SuperPCA_pVals() instead