Calculate the Optimal Parameters for a Mixture of Weibull Distributions

Calculate the parameters which minimise the likelihood from a mixture of two Weibull Extreme Value distributions.

weibullMix_optimParams(max_tControl_vec, pathwaySize_vec, initialVals = c(p =
  0.5, mu1 = 1, s1 = 0.5, mu2 = 1, s2 = 0.5), optimMethod = "L-BFGS-B",
  lowerBD = c(0, -Inf, 0, -Inf, 0), upperBD = c(1, Inf, Inf, Inf, Inf))

Arguments

max_tControl_vec
A vector of the maximum absolute t-scores for each pathway when under the null model (the response vector has been randomly assigned or parametrically bootstrapped).

pathwaySize_vec
A vector of the number of genes in each pathway

initialVals
A vector of initial values for the Weibull parameters. The values are

p : The mixing proportion between the Gumbel minimum and Gumbel maximum distributions. Defaults to 0.5.

mu1 : The mean of the first distribution. Defaults to 1.

s1 : The standard deviation (precision?) of the first distribution. Defaults to 0.5.

mu2 : The mean of the second distribution. Defaults to 1.

s2 : The standard deviation (precision?) of the second distribution. Defaults to 0.5.

optimMethod
Which numerical optimization routine to pass to the optim function. Defaults to "L-BFGS-B", which allows for lower and upper bound constraints. When this option is selected, lower and upper bounds for ALL parameters must be supplied.

lowerBD
A vector of the lower bounds on the initialVals.

upperBD
A vector of the upper bounds on the initialVals.

Value

A named vector of the estimated values for the parameters which minimize the likelihood.

Details

The likelihood function is equation (4) Chen et al (2008): a mixture of two Gumbel Extreme Value pdfs, with mixing proportion p. The values mu_i and s_i, i = 1, 2, within the code of the function are placeholders for the mean and standard deviation, respectively.

A computational note: the "L-BFGS-B" option within the optim function requires a bounded function or likelihood. We therefore replaced Inf with $10 ^ 200$ in the check for boundedness. As we are attempting to minimise the likelihood, this maximum machine value is effectively Infinity.

See https://doi.org/10.1093/bioinformatics/btn458 for more information

@seealso optim

Examples

   NULL

#> NULL